feat: Prevent changes unrelated to a proof from changing the verification behavior of the proof (#1612)
In DafnyOptions.cs, turn on Boogie's `Prune` and `NormaliseNames` option, and turn off `EmitDebugInformation`, to increase verification stability. Use `uses` clauses on constants and functions, both in DafnyPrelude.bpl and in generated Boogie declarations, and the `include_dep` attribute on axioms, to ensure not too much is pruned. More information on these Boogie features can be found in the PRs that introduce them: https://github.com/boogie-org/boogie/pull/450, https://github.com/boogie-org/boogie/pull/452, https://github.com/boogie-org/boogie/pull/453, https://github.com/boogie-org/boogie/pull/454, https://github.com/boogie-org/boogie/pull/464

In follow-up PRs, we will replace all `include_dep` with `uses` clauses, since they allow more but still correct pruning.

This PR also updates the test `SchorrWaite.dfy` to make it more stable. Changes stolen from https://github.com/dafny-lang/dafny/pull/1620
