+++ Source/Dafny/Verifier/Translator.cs
@@ -22,6 +22,26 @@ namespace Microsoft.Dafny {
     // TODO(wuestholz): Enable this once Dafny's recommended Z3 version includes changeset 0592e765744497a089c42021990740f303901e67.
     public bool UseOptimizationInZ3 { get; set; }
 
+    void AddOtherDefinition(Bpl.Declaration declaration, Axiom axiom) {
+
+      switch (declaration) {
+        case Boogie.Function boogieFunction:
+          boogieFunction.AddOtherDefinitionAxiom(axiom);
+          break;
+        case Boogie.Constant boogieConstant:
+          boogieConstant.DefinitionAxioms.Add(axiom);
+          break;
+        default: throw new ArgumentException("Declaration must be a function or constant");
+      }
+
+      sink.AddTopLevelDeclaration(axiom);
+    }
+
+    void AddIncludeDepAxiom(Axiom axiom) {
+      axiom.AddAttribute("include_dep");
+      sink.AddTopLevelDeclaration(axiom);
+    }
+
     public class TranslatorFlags {
       public bool InsertChecksums = 0 < CommandLineOptions.Clo.VerifySnapshots;
       public string UniqueIdPrefix = null;
@@ -186,6 +206,8 @@ namespace Microsoft.Dafny {
       public readonly Bpl.Function IMapValues;
       public readonly Bpl.Function MapItems;
       public readonly Bpl.Function IMapItems;
+      public readonly Bpl.Function ObjectTypeConstructor;
+      public readonly Bpl.Function Tuple2TypeConstructor;
       public readonly Bpl.Function Tuple2Destructors0;
       public readonly Bpl.Function Tuple2Destructors1;
       public readonly Bpl.Function Tuple2Constructor;
@@ -296,7 +318,8 @@ namespace Microsoft.Dafny {
                              Bpl.Function ORD_isLimit, Bpl.Function ORD_isSucc, Bpl.Function ORD_offset, Bpl.Function ORD_isNat,
                              Bpl.Function mapDomain, Bpl.Function imapDomain,
                              Bpl.Function mapValues, Bpl.Function imapValues, Bpl.Function mapItems, Bpl.Function imapItems,
-                             Bpl.Function tuple2Destructors0, Bpl.Function tuple2Destructors1, Bpl.Function tuple2Constructor,
+                             Bpl.Function objectTypeConstructor,
+                             Bpl.Function tuple2Destructors0, Bpl.Function tuple2Destructors1, Bpl.Function tuple2Constructor, Bpl.Function tuple2TypeConstructor,
                              Bpl.TypeCtorDecl seqTypeCtor, Bpl.TypeSynonymDecl bv0TypeDecl,
                              Bpl.TypeCtorDecl fieldNameType, Bpl.TypeCtorDecl tyType, Bpl.TypeCtorDecl tyTagType, Bpl.TypeCtorDecl tyTagFamilyType,
                              Bpl.GlobalVariable heap, Bpl.TypeCtorDecl classNameType, Bpl.TypeCtorDecl nameFamilyType,
@@ -363,9 +386,11 @@ namespace Microsoft.Dafny {
         this.IMapValues = imapValues;
         this.MapItems = mapItems;
         this.IMapItems = imapItems;
+        this.ObjectTypeConstructor = objectTypeConstructor;
         this.Tuple2Destructors0 = tuple2Destructors0;
         this.Tuple2Destructors1 = tuple2Destructors1;
         this.Tuple2Constructor = tuple2Constructor;
+        this.Tuple2TypeConstructor = tuple2TypeConstructor;
         this.seqTypeCtor = seqTypeCtor;
         this.Bv0Type = new Bpl.TypeSynonymAnnotation(Token.NoToken, bv0TypeDecl, new List<Bpl.Type>());
         this.fieldName = fieldNameType;
@@ -409,6 +434,8 @@ namespace Microsoft.Dafny {
       Bpl.Function imapValues = null;
       Bpl.Function mapItems = null;
       Bpl.Function imapItems = null;
+      Bpl.Function objectTypeConstructor = null;
+      Bpl.Function tuple2TypeConstructor = null;
       Bpl.Function tuple2Destructors0 = null;
       Bpl.Function tuple2Destructors1 = null;
       Bpl.Function tuple2Constructor = null;
@@ -521,6 +548,10 @@ namespace Microsoft.Dafny {
             tuple2Destructors1 = f;
           } else if (f.Name == "#_System._tuple#2._#Make2") {
             tuple2Constructor = f;
+          } else if (f.Name == "Tclass._System.Tuple2") {
+            tuple2TypeConstructor = f;
+          } else if (f.Name == "Tclass._System.object?") {
+            objectTypeConstructor = f;
           }
         }
       }
@@ -600,6 +631,10 @@ namespace Microsoft.Dafny {
         Console.WriteLine("Error: Dafny prelude is missing declaration of $Heap");
       } else if (allocField == null) {
         Console.WriteLine("Error: Dafny prelude is missing declaration of constant alloc");
+      } else if (tuple2TypeConstructor == null) {
+        Console.WriteLine("Error: Dafny prelude is missing declaration of tuple2TypeConstructor");
+      } else if (objectTypeConstructor == null) {
+        Console.WriteLine("Error: Dafny prelude is missing declaration of objectTypeConstructor");
       } else {
         return new PredefinedDecls(charType, refType, boxType, tickType,
                                    setTypeCtor, isetTypeCtor, multiSetTypeCtor,
@@ -608,7 +643,8 @@ namespace Microsoft.Dafny {
                                    ORDINAL_isLimit, ORDINAL_isSucc, ORDINAL_offset, ORDINAL_isNat,
                                    mapDomain, imapDomain,
                                    mapValues, imapValues, mapItems, imapItems,
-                                   tuple2Destructors0, tuple2Destructors1, tuple2Constructor,
+                                   objectTypeConstructor,
+                                   tuple2Destructors0, tuple2Destructors1, tuple2Constructor, tuple2TypeConstructor,
                                    seqTypeCtor, bv0TypeDecl,
                                    fieldNameType, tyType, tyTagType, tyTagFamilyType,
                                    heap, classNameType, nameFamilyType,
@@ -1053,6 +1089,7 @@ namespace Microsoft.Dafny {
             bvs.Add(oVar);
             var tr = BplTrigger(isC);
             var body = BplImp(BplAnd(oNotNull, isC), isJ);
+
             sink.AddTopLevelDeclaration(new Bpl.Axiom(c.tok, new Bpl.ForallExpr(c.tok, bvs, tr, body)));
 
             // axiom (forall T: Ty, $Heap: Heap, $o: ref ::
@@ -1205,6 +1242,7 @@ namespace Microsoft.Dafny {
       currentModule = null;
       this.fuelContext = oldFuelContext;
     }
+
     void AddRedirectingTypeDeclAxioms<T>(bool is_alloc, T dd, string fullName) where T : TopLevelDecl, RedirectingTypeDecl {
       Contract.Requires(dd != null);
       Contract.Requires(dd.Var != null && dd.Constraint != null);
@@ -1251,664 +1289,10 @@ namespace Microsoft.Dafny {
         body = BplIff(is_o, BplAnd(parentConstraint, constraint));
       }
 
-      sink.AddTopLevelDeclaration(new Bpl.Axiom(dd.tok, BplForall(vars, BplTrigger(is_o), body), name));
+      AddIncludeDepAxiom(new Bpl.Axiom(dd.tok, BplForall(vars, BplTrigger(is_o), body), name));
     }
 
-    void AddDatatype(DatatypeDecl dt) {
-      Contract.Requires(dt != null);
-      Contract.Requires(sink != null && predef != null);
 
-      foreach (DatatypeCtor ctor in dt.Ctors) {
-        // Add:  function #dt.ctor(tyVars, paramTypes) returns (DatatypeType);
-
-        List<Bpl.Variable> argTypes = new List<Bpl.Variable>();
-        foreach (Formal arg in ctor.Formals) {
-          Bpl.Variable a = new Bpl.Formal(arg.tok, new Bpl.TypedIdent(arg.tok, Bpl.TypedIdent.NoName, TrType(arg.Type)), true);
-          argTypes.Add(a);
-        }
-        Bpl.Variable resType = new Bpl.Formal(ctor.tok, new Bpl.TypedIdent(ctor.tok, Bpl.TypedIdent.NoName, predef.DatatypeType), false);
-        Bpl.Function fn;
-        if (dt is TupleTypeDecl ttd && ttd.Dims == 2 && ttd.NonGhostDims == 2) {
-          fn = predef.Tuple2Constructor;
-        } else {
-          fn = new Bpl.Function(ctor.tok, ctor.FullName, argTypes, resType, "Constructor function declaration");
-          sink.AddTopLevelDeclaration(fn);
-        }
-        if (InsertChecksums) {
-          InsertChecksum(dt, fn);
-        }
-
-        List<Bpl.Variable> bvs;
-        List<Bpl.Expr> args;
-
-
-        {
-          // Add:  const unique ##dt.ctor: DtCtorId;
-          Bpl.Constant cid = new Bpl.Constant(ctor.tok, new Bpl.TypedIdent(ctor.tok, "#" + ctor.FullName, predef.DtCtorId), true);
-          Bpl.Expr c = new Bpl.IdentifierExpr(ctor.tok, cid);
-          sink.AddTopLevelDeclaration(cid);
-
-          {
-            // Add:  axiom (forall params :: DatatypeCtorId(#dt.ctor(params)) == ##dt.ctor);
-            CreateBoundVariables(ctor.Formals, out bvs, out args);
-            var constructor_call = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-            var lhs = FunctionCall(ctor.tok, BuiltinFunction.DatatypeCtorId, null, constructor_call);
-            Bpl.Expr q = Bpl.Expr.Eq(lhs, c);
-            var trigger = BplTrigger(constructor_call);
-            sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, BplForall(bvs, trigger, q), "Constructor identifier"));
-          }
-
-          {
-            // Add:  function dt.ctor?(this: DatatypeType): bool { DatatypeCtorId(this) == ##dt.ctor }
-            fn = GetReadonlyField(ctor.QueryField);
-            sink.AddTopLevelDeclaration(fn);
-
-            // and here comes the associated axiom:
-
-            Bpl.Expr th; var thVar = BplBoundVar("d", predef.DatatypeType, out th);
-            var queryPredicate = FunctionCall(ctor.tok, fn.Name, Bpl.Type.Bool, th);
-            var ctorId = FunctionCall(ctor.tok, BuiltinFunction.DatatypeCtorId, null, th);
-            var rhs = Bpl.Expr.Eq(ctorId, c);
-            var body = Bpl.Expr.Iff(queryPredicate, rhs);
-            var tr = BplTrigger(queryPredicate);
-            var ax = BplForall(thVar, tr, body);
-            sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, ax, "Questionmark and identifier"));
-          }
-
-          // check well-formedness of any default-value expressions
-          AddWellformednessCheck(ctor);
-        }
-
-
-        {
-          // Add:  axiom (forall d: DatatypeType :: dt.ctor?(d) ==> (exists params :: d == #dt.ctor(params));
-          CreateBoundVariables(ctor.Formals, out bvs, out args);
-          Bpl.Expr rhs = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-          Bpl.Expr dId; var dBv = BplBoundVar("d", predef.DatatypeType, out dId);
-          Bpl.Expr q = Bpl.Expr.Eq(dId, rhs);
-          if (bvs.Count != 0) {
-            q = new Bpl.ExistsExpr(ctor.tok, bvs, null/*always in a Skolemization context*/, q);
-          }
-          Bpl.Expr dtq = FunctionCall(ctor.tok, ctor.QueryField.FullSanitizedName, Bpl.Type.Bool, dId);
-          var trigger = BplTrigger(dtq);
-          q = BplForall(dBv, trigger, BplImp(dtq, q));
-          sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, q, "Constructor questionmark has arguments"));
-        }
-
-        MapM(Bools, is_alloc => {
-          /*
-            (forall x0 : C0, ..., xn : Cn, G : Ty •
-              { $Is(C(x0,...,xn), T(G)) }
-              $Is(C(x0,...,xn), T(G)) <==>
-              $Is[Box](x0, C0(G)) && ... && $Is[Box](xn, Cn(G)));
-            (forall x0 : C0, ..., xn : Cn, G : Ty, H : Heap •
-                { $IsAlloc(C(G, x0,...,xn), T(G), H) }
-                IsGoodHeap(H) ==>
-                   ($IsAlloc(C(G, x0,...,xn), T(G), H) <==>
-                    $IsAlloc[Box](x0, C0(G), H) && ... && $IsAlloc[Box](xn, Cn(G), H)));
-          */
-          List<Bpl.Expr> tyexprs;
-          var tyvars = MkTyParamBinders(dt.TypeArgs, out tyexprs);
-          CreateBoundVariables(ctor.Formals, out bvs, out args);
-          Bpl.Expr h;
-          var hVar = BplBoundVar("$h", predef.HeapType, out h);
-          Bpl.Expr conj = Bpl.Expr.True;
-          for (var i = 0; i < ctor.Formals.Count; i++) {
-            var arg = ctor.Formals[i];
-            if (is_alloc) {
-              if (CommonHeapUse || (NonGhostsUseHeap && !arg.IsGhost)) {
-                conj = BplAnd(conj, MkIsAlloc(args[i], arg.Type, h));
-              }
-            } else {
-              conj = BplAnd(conj, MkIs(args[i], arg.Type));
-            }
-          }
-          var c_params = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-          var c_ty = ClassTyCon((TopLevelDecl)dt, tyexprs);
-          bvs.InsertRange(0, tyvars);
-          if (!is_alloc) {
-            var c_is = MkIs(c_params, c_ty);
-            sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok,
-                BplForall(bvs, BplTrigger(c_is), BplIff(c_is, conj)),
-                "Constructor $Is"));
-          } else if (is_alloc && (CommonHeapUse || NonGhostsUseHeap)) {
-            var isGoodHeap = FunctionCall(ctor.tok, BuiltinFunction.IsGoodHeap, null, h);
-            var c_alloc = MkIsAlloc(c_params, c_ty, h);
-            bvs.Add(hVar);
-            sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok,
-                BplForall(bvs, BplTrigger(c_alloc),
-                               BplImp(isGoodHeap, BplIff(c_alloc, conj))),
-                "Constructor $IsAlloc"));
-          }
-          if (is_alloc && CommonHeapUse && !AlwaysUseHeap) {
-            for (int i = 0; i < ctor.Formals.Count; i++) {
-              var arg = ctor.Formals[i];
-              var dtor = GetReadonlyField(ctor.Destructors[i]);
-              /* (forall d : DatatypeType, G : Ty, H : Heap •
-                     { $IsAlloc[Box](Dtor(d), D(G), H) }
-                     IsGoodHeap(H) &&
-                     C?(d) &&
-                     (exists G' : Ty :: $IsAlloc(d, T(G,G'), H))
-                     ==>
-                         $IsAlloc[Box](Dtor(d), D(G), H))
-               */
-              Bpl.Expr dId; var dBv = BplBoundVar("d", predef.DatatypeType, out dId);
-              var isGoodHeap = FunctionCall(ctor.tok, BuiltinFunction.IsGoodHeap, null, h);
-              Bpl.Expr dtq = FunctionCall(ctor.tok, ctor.QueryField.FullSanitizedName, Bpl.Type.Bool, dId);
-              var c_alloc = MkIsAlloc(dId, c_ty, h);
-              var dtorD = FunctionCall(ctor.tok, dtor.Name, TrType(arg.Type), dId);
-              var d_alloc = MkIsAlloc(dtorD, arg.Type, h);
-
-              // split tyvars into G,G' where G are the type variables that are used in the type of the destructor
-              var freeTypeVars = new HashSet<TypeParameter>();
-              ComputeFreeTypeVariables_All(arg.Type, freeTypeVars);
-              var tyvarsG = new List<Bpl.Variable>();
-              var tyvarsGprime = new List<Bpl.Variable>();
-              Contract.Assert(dt.TypeArgs.Count == tyvars.Count);
-              for (int j = 0; j < dt.TypeArgs.Count; j++) {
-                var tv = tyvars[j];
-                if (freeTypeVars.Contains(dt.TypeArgs[j])) {
-                  tyvarsG.Add(tv);
-                } else {
-                  tyvarsGprime.Add(tv);
-                }
-              }
-
-              bvs = new List<Bpl.Variable>();
-              bvs.Add(dBv);
-              bvs.AddRange(tyvarsG);
-              bvs.Add(hVar);
-              if (tyvarsGprime.Count != 0) {
-                c_alloc = new Bpl.ExistsExpr(ctor.tok, tyvarsGprime, BplTrigger(c_alloc), c_alloc);
-              }
-              sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok,
-                  BplForall(bvs, BplTrigger(d_alloc),
-                                 BplImp(BplAnd(isGoodHeap, BplAnd(dtq, c_alloc)), d_alloc)),
-                  "Destructor $IsAlloc"));
-            }
-          }
-        });
-
-        if (dt is IndDatatypeDecl) {
-          // Add Lit axiom:
-          // axiom (forall p0, ..., pn :: #dt.ctor(Lit(p0), ..., Lit(pn)) == Lit(#dt.ctor(p0, .., pn)));
-          CreateBoundVariables(ctor.Formals, out bvs, out args);
-          var litargs = new List<Bpl.Expr>();
-          foreach (Bpl.Expr arg in args) {
-            litargs.Add(Lit(arg));
-          }
-          Bpl.Expr lhs = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, litargs);
-          Bpl.Expr rhs = Lit(FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args), predef.DatatypeType);
-          Bpl.Expr q = BplForall(bvs, BplTrigger(lhs), Bpl.Expr.Eq(lhs, rhs));
-          sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, q, "Constructor literal"));
-        }
-
-        // Injectivity axioms for normal arguments
-        for (int i = 0; i < ctor.Formals.Count; i++) {
-          var arg = ctor.Formals[i];
-          // function ##dt.ctor#i(DatatypeType) returns (Ti);
-          var sf = ctor.Destructors[i];
-          Contract.Assert(sf != null);
-          fn = GetReadonlyField(sf);
-          if (fn == predef.Tuple2Destructors0 || fn == predef.Tuple2Destructors1) {
-            // the two destructors for 2-tuples are predefined in Prelude for use
-            // by the Map#Items axiom
-          } else if (sf.EnclosingCtors[0] != ctor) {
-            // this special field, which comes from a shared destructor, is being declared in a different iteration of this loop
-          } else {
-            sink.AddTopLevelDeclaration(fn);
-          }
-          // axiom (forall params :: ##dt.ctor#i(#dt.ctor(params)) == params_i);
-          CreateBoundVariables(ctor.Formals, out bvs, out args);
-          var inner = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-          var outer = FunctionCall(ctor.tok, fn.Name, TrType(arg.Type), inner);
-          var q = BplForall(bvs, BplTrigger(inner), Bpl.Expr.Eq(outer, args[i]));
-          sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, q, "Constructor injectivity"));
-
-          if (dt is IndDatatypeDecl) {
-            var argType = arg.Type.NormalizeExpandKeepConstraints();  // TODO: keep constraints -- really?  Write a test case
-            if (argType.IsDatatype || argType.IsTypeParameter) {
-              // for datatype:             axiom (forall params :: {#dt.ctor(params)} DtRank(params_i) < DtRank(#dt.ctor(params)));
-              // for type-parameter type:  axiom (forall params :: {#dt.ctor(params)} BoxRank(params_i) < DtRank(#dt.ctor(params)));
-              CreateBoundVariables(ctor.Formals, out bvs, out args);
-              Bpl.Expr lhs = FunctionCall(ctor.tok, arg.Type.IsDatatype ? BuiltinFunction.DtRank : BuiltinFunction.BoxRank, null, args[i]);
-              /* CHECK
-              Bpl.Expr lhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null,
-                argType.IsDatatype ? args[i] : FunctionCall(ctor.tok, BuiltinFunction.Unbox, predef.DatatypeType, args[i]));
-              */
-              Bpl.Expr ct = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-              var rhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, ct);
-              var trigger = BplTrigger(ct);
-              q = new Bpl.ForallExpr(ctor.tok, bvs, trigger, Bpl.Expr.Lt(lhs, rhs));
-              sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, q, "Inductive rank"));
-            } else if (argType is SeqType) {
-              // axiom (forall params, i: int {#dt.ctor(params)} :: 0 <= i && i < |arg| ==> DtRank(arg[i]) < DtRank(#dt.ctor(params)));
-              // that is:
-              // axiom (forall params, i: int {#dt.ctor(params)} :: 0 <= i && i < |arg| ==> DtRank(Unbox(Seq#Index(arg,i))) < DtRank(#dt.ctor(params)));
-              {
-                CreateBoundVariables(ctor.Formals, out bvs, out args);
-                Bpl.Variable iVar = new Bpl.BoundVariable(arg.tok, new Bpl.TypedIdent(arg.tok, "i", Bpl.Type.Int));
-                bvs.Add(iVar);
-                Bpl.IdentifierExpr ie = new Bpl.IdentifierExpr(arg.tok, iVar);
-                Bpl.Expr ante = Bpl.Expr.And(
-                  Bpl.Expr.Le(Bpl.Expr.Literal(0), ie),
-                  Bpl.Expr.Lt(ie, FunctionCall(arg.tok, BuiltinFunction.SeqLength, null, args[i])));
-                var seqIndex = FunctionCall(arg.tok, BuiltinFunction.SeqIndex, predef.DatatypeType, args[i], ie);
-                Bpl.Expr lhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null,
-                  FunctionCall(arg.tok, BuiltinFunction.Unbox, predef.DatatypeType, seqIndex));
-                var ct = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-                var rhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, ct);
-                q = new Bpl.ForallExpr(ctor.tok, bvs, new Trigger(lhs.tok, true, new List<Bpl.Expr> { seqIndex, ct }), Bpl.Expr.Imp(ante, Bpl.Expr.Lt(lhs, rhs)));
-                sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, q, "Inductive seq element rank"));
-              }
-
-              // axiom (forall params {#dt.ctor(params)} :: SeqRank(arg) < DtRank(#dt.ctor(params)));
-              {
-                CreateBoundVariables(ctor.Formals, out bvs, out args);
-                var lhs = FunctionCall(ctor.tok, BuiltinFunction.SeqRank, null, args[i]);
-                var ct = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-                var rhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, ct);
-                var trigger = BplTrigger(ct);
-                q = new Bpl.ForallExpr(ctor.tok, bvs, trigger, Bpl.Expr.Lt(lhs, rhs));
-                sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, q, "Inductive seq rank"));
-              }
-            } else if (argType is SetType) {
-              // axiom (forall params, d: Datatype {arg[d], #dt.ctor(params)}  :: arg[d] ==> DtRank(d) < DtRank(#dt.ctor(params)));
-              // that is:
-              // axiom (forall params, d: Datatype {arg[Box(d)], #dt.ctor(params)} :: arg[Box(d)] ==> DtRank(d) < DtRank(#dt.ctor(params)));
-              CreateBoundVariables(ctor.Formals, out bvs, out args);
-              Bpl.Variable dVar = new Bpl.BoundVariable(arg.tok, new Bpl.TypedIdent(arg.tok, "d", predef.DatatypeType));
-              bvs.Add(dVar);
-              Bpl.IdentifierExpr ie = new Bpl.IdentifierExpr(arg.tok, dVar);
-              Bpl.Expr inSet = Bpl.Expr.SelectTok(arg.tok, args[i], FunctionCall(arg.tok, BuiltinFunction.Box, null, ie));
-              Bpl.Expr lhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, ie);
-              var ct = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-              var rhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, ct);
-              var trigger = new Bpl.Trigger(ctor.tok, true, new List<Bpl.Expr> { inSet, ct });
-              q = new Bpl.ForallExpr(ctor.tok, bvs, trigger, Bpl.Expr.Imp(inSet, Bpl.Expr.Lt(lhs, rhs)));
-              sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, q, "Inductive set element rank"));
-            } else if (argType is MultiSetType) {
-              // axiom (forall params, d: Datatype {arg[d], #dt.ctor(params)} :: 0 < arg[d] ==> DtRank(d) < DtRank(#dt.ctor(params)));
-              // that is:
-              // axiom (forall params, d: Datatype {arg[Box(d)], #dt.ctor(params)} :: 0 < arg[Box(d)] ==> DtRank(d) < DtRank(#dt.ctor(params)));
-              CreateBoundVariables(ctor.Formals, out bvs, out args);
-              Bpl.Variable dVar = new Bpl.BoundVariable(arg.tok, new Bpl.TypedIdent(arg.tok, "d", predef.DatatypeType));
-              bvs.Add(dVar);
-              Bpl.IdentifierExpr ie = new Bpl.IdentifierExpr(arg.tok, dVar);
-              var inMultiset = Bpl.Expr.SelectTok(arg.tok, args[i], FunctionCall(arg.tok, BuiltinFunction.Box, null, ie));
-              Bpl.Expr ante = Bpl.Expr.Gt(inMultiset, Bpl.Expr.Literal(0));
-              Bpl.Expr lhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, ie);
-              var ct = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-              var rhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, ct);
-              var trigger = new Bpl.Trigger(ctor.tok, true, new List<Bpl.Expr> { inMultiset, ct });
-              q = new Bpl.ForallExpr(ctor.tok, bvs, trigger, Bpl.Expr.Imp(ante, Bpl.Expr.Lt(lhs, rhs)));
-              sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, q, "Inductive multiset element rank"));
-            } else if (argType is MapType) {
-              var finite = ((MapType)argType).Finite;
-              {
-                // axiom (forall params, d: DatatypeType
-                //   { Map#Domain(arg)[$Box(d)], #dt.ctor(params) }
-                //   Map#Domain(arg)[$Box(d)] ==> DtRank(d) < DtRank(#dt.ctor(params)));
-                CreateBoundVariables(ctor.Formals, out bvs, out args);
-                var dVar = new Bpl.BoundVariable(arg.tok, new Bpl.TypedIdent(arg.tok, "d", predef.DatatypeType));
-                bvs.Add(dVar);
-                var ie = new Bpl.IdentifierExpr(arg.tok, dVar);
-                var f = finite ? BuiltinFunction.MapDomain : BuiltinFunction.IMapDomain;
-                var domain = FunctionCall(arg.tok, f, predef.MapType(arg.tok, finite, predef.BoxType, predef.BoxType), args[i]);
-                var inDomain = Bpl.Expr.SelectTok(arg.tok, domain, FunctionCall(arg.tok, BuiltinFunction.Box, null, ie));
-                var lhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, ie);
-                var ct = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-                var rhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, ct);
-                var trigger = new Bpl.Trigger(ctor.tok, true, new List<Bpl.Expr> { inDomain, ct });
-                q = new Bpl.ForallExpr(ctor.tok, bvs, trigger, Bpl.Expr.Imp(inDomain, Bpl.Expr.Lt(lhs, rhs)));
-                sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, q, "Inductive map key rank"));
-              }
-              {
-                // axiom(forall params, bx: Box ::
-                //   { Map#Elements(arg)[bx], #dt.ctor(params) }
-                //   Map#Domain(arg)[bx] ==> DtRank($Unbox(Map#Elements(arg)[bx]): DatatypeType) < DtRank(#dt.ctor(params)));
-                CreateBoundVariables(ctor.Formals, out bvs, out args);
-                var bxVar = new Bpl.BoundVariable(arg.tok, new Bpl.TypedIdent(arg.tok, "bx", predef.BoxType));
-                bvs.Add(bxVar);
-                var ie = new Bpl.IdentifierExpr(arg.tok, bxVar);
-                var f = finite ? BuiltinFunction.MapDomain : BuiltinFunction.IMapDomain;
-                var domain = FunctionCall(arg.tok, f, predef.MapType(arg.tok, finite, predef.BoxType, predef.BoxType), args[i]);
-                var inDomain = Bpl.Expr.SelectTok(arg.tok, domain, ie);
-                var ef = finite ? BuiltinFunction.MapElements : BuiltinFunction.IMapElements;
-                var element = FunctionCall(arg.tok, ef, predef.MapType(arg.tok, finite, predef.BoxType, predef.BoxType), args[i]);
-                var elmt = Bpl.Expr.SelectTok(arg.tok, element, ie);
-                var unboxElmt = FunctionCall(arg.tok, BuiltinFunction.Unbox, predef.DatatypeType, elmt);
-                var lhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, unboxElmt);
-                var ct = FunctionCall(ctor.tok, ctor.FullName, predef.DatatypeType, args);
-                var rhs = FunctionCall(ctor.tok, BuiltinFunction.DtRank, null, ct);
-                var trigger = new Bpl.Trigger(ctor.tok, true, new List<Bpl.Expr> { inDomain, ct });
-                q = new Bpl.ForallExpr(ctor.tok, bvs, trigger, Bpl.Expr.Imp(inDomain, Bpl.Expr.Lt(lhs, rhs)));
-                sink.AddTopLevelDeclaration(new Bpl.Axiom(ctor.tok, q, "Inductive map value rank"));
-              }
-            }
-          }
-        }
-      }
-
-      {
-        // Add:
-        //   function $IsA#Dt(G: Ty,d: DatatypeType): bool {
-        //     Dt.Ctor0?(G, d) || Dt.Ctor1?(G, d) || ...
-        //   }
-        var cases_dBv = new Bpl.Formal(dt.tok, new Bpl.TypedIdent(dt.tok, Bpl.TypedIdent.NoName, predef.DatatypeType), true);
-        var cases_resType = new Bpl.Formal(dt.tok, new Bpl.TypedIdent(dt.tok, Bpl.TypedIdent.NoName, Bpl.Type.Bool), false);
-        var cases_fn = new Bpl.Function(dt.tok, "$IsA#" + dt.FullSanitizedName,
-                                        new List<Variable> { cases_dBv },
-                                        cases_resType,
-                                        "Depth-one case-split function");
-
-        if (InsertChecksums) {
-          InsertChecksum(dt, cases_fn);
-        }
-
-        sink.AddTopLevelDeclaration(cases_fn);
-        // and here comes the actual axiom:
-        {
-          Bpl.Expr d;
-          var dVar = BplBoundVar("d", predef.DatatypeType, out d);
-          var lhs = FunctionCall(dt.tok, cases_fn.Name, Bpl.Type.Bool, d);
-          Bpl.Expr cases_body = Bpl.Expr.False;
-          foreach (DatatypeCtor ctor in dt.Ctors) {
-            var disj = FunctionCall(ctor.tok, ctor.QueryField.FullSanitizedName, Bpl.Type.Bool, d);
-            cases_body = BplOr(cases_body, disj);
-          }
-          var ax = BplForall(new List<Variable> { dVar }, BplTrigger(lhs), BplImp(lhs, cases_body));
-          sink.AddTopLevelDeclaration(new Bpl.Axiom(dt.tok, ax, "Depth-one case-split axiom"));
-        }
-      }
-
-      // The axiom above ($IsA#Dt(d) <==> Dt.Ctor0?(d) || Dt.Ctor1?(d)) gets triggered only with $IsA#Dt(d).  The $IsA#Dt(d)
-      // predicate is generated only where the translation inserts it; in other words, the user cannot write any assertion
-      // that causes the $IsA#Dt(d) predicate to be emitted.  This is what we want, because making the RHS disjunction be
-      // available too often makes performance go down.  However, we do want to allow the disjunction to be introduced if the
-      // user explicitly talks about one of its disjuncts.  To make this useful, we introduce the following axiom.  Note that
-      // the DtType(d) information is available everywhere.
-      // axiom (forall G: Ty, d: DatatypeType ::
-      //         { Dt.Ctor0?(G,d) }
-      //         { Dt.Ctor1?(G,d) }
-      //         $Is(d, T(G)) ==> Dt.Ctor0?(G,d) || Dt.Ctor1?(G,d) || ...);
-      {
-        List<Bpl.Expr> tyexprs;
-        var tyvars = MkTyParamBinders(dt.TypeArgs, out tyexprs);
-        Bpl.Expr d;
-        var dVar = BplBoundVar("d", predef.DatatypeType, out d);
-        var d_is = MkIs(d, ClassTyCon(dt, tyexprs));
-        Bpl.Expr cases_body = Bpl.Expr.False;
-        Bpl.Trigger tr = null;
-        foreach (DatatypeCtor ctor in dt.Ctors) {
-          var disj = FunctionCall(ctor.tok, ctor.QueryField.FullSanitizedName, Bpl.Type.Bool, d);
-          cases_body = BplOr(cases_body, disj);
-          tr = new Bpl.Trigger(ctor.tok, true, new List<Bpl.Expr> { disj, d_is }, tr);
-        }
-        var body = Bpl.Expr.Imp(d_is, cases_body);
-        var ax = BplForall(Snoc(tyvars, dVar), tr, body);
-        sink.AddTopLevelDeclaration(new Bpl.Axiom(dt.tok, ax, "Questionmark data type disjunctivity"));
-      }
-
-      if (dt is IndDatatypeDecl) {
-        var dtEqualName = dt.FullSanitizedName + "#Equal";
-
-        // Add function Dt#Equal(DatatypeType, DatatypeType): bool;
-        // For each constructor Ctor(x: X, y: Y), add an axiom of the form
-        //     forall a, b ::
-        //       { Dt#Equal(a, b), Ctor?(a) }
-        //       { Dt#Equal(a, b), Ctor?(b) }
-        //       Ctor?(a) && Ctor?(b)
-        //       ==>
-        //       (Dt#Equal(a, b) <==>
-        //           X#Equal(a.x, b.x) &&
-        //           Y#Equal(a.y, b.y)
-        //       )
-        // where X#Equal is the equality predicate for type X and a.x denotes Dtor#x(a), and similarly
-        // for Y and b.
-        // Except, in the event that the datatype has exactly one constructor, then instead generate:
-        //     forall a, b ::
-        //       { Dt#Equal(a, b) }
-        //       true
-        //       ==>
-        //       ...as before
-        {
-          var args = new List<Variable>();
-          args.Add(new Bpl.Formal(dt.tok, new Bpl.TypedIdent(dt.tok, Bpl.TypedIdent.NoName, predef.DatatypeType), false));
-          args.Add(new Bpl.Formal(dt.tok, new Bpl.TypedIdent(dt.tok, Bpl.TypedIdent.NoName, predef.DatatypeType), false));
-          var ctorEqualResult = new Bpl.Formal(dt.tok, new Bpl.TypedIdent(dt.tok, Bpl.TypedIdent.NoName, Bpl.Type.Bool), false);
-          sink.AddTopLevelDeclaration(new Bpl.Function(dt.tok, dtEqualName, args, ctorEqualResult, "Datatype extensional equality declaration"));
-
-          Bpl.Expr a; var aVar = BplBoundVar("a", predef.DatatypeType, out a);
-          Bpl.Expr b; var bVar = BplBoundVar("b", predef.DatatypeType, out b);
-
-          var dtEqual = FunctionCall(dt.tok, dtEqualName, Bpl.Type.Bool, a, b);
-
-          foreach (var ctor in dt.Ctors) {
-            Bpl.Trigger trigger;
-            Bpl.Expr ante;
-            if (dt.Ctors.Count == 1) {
-              ante = Bpl.Expr.True;
-              trigger = BplTrigger(dtEqual);
-            } else {
-              var ctorQ = GetReadonlyField(ctor.QueryField);
-              var ctorQa = FunctionCall(ctor.tok, ctorQ.Name, Bpl.Type.Bool, a);
-              var ctorQb = FunctionCall(ctor.tok, ctorQ.Name, Bpl.Type.Bool, b);
-              ante = BplAnd(ctorQa, ctorQb);
-              trigger = dt.Ctors.Count == 1 ? BplTrigger(dtEqual) :
-                new Bpl.Trigger(ctor.tok, true, new List<Bpl.Expr> { dtEqual, ctorQa },
-                new Bpl.Trigger(ctor.tok, true, new List<Bpl.Expr> { dtEqual, ctorQb }));
-            }
-
-            Bpl.Expr eqs = Bpl.Expr.True;
-            for (var i = 0; i < ctor.Formals.Count; i++) {
-              var arg = ctor.Formals[i];
-              var dtor = GetReadonlyField(ctor.Destructors[i]);
-              var dtorA = FunctionCall(ctor.tok, dtor.Name, TrType(arg.Type), a);
-              var dtorB = FunctionCall(ctor.tok, dtor.Name, TrType(arg.Type), b);
-              var eq = TypeSpecificEqual(ctor.tok, arg.Type, dtorA, dtorB);
-              eqs = BplAnd(eqs, eq);
-            }
-
-            var ax = BplForall(new List<Variable> { aVar, bVar }, trigger, Bpl.Expr.Imp(ante, Bpl.Expr.Iff(dtEqual, eqs)));
-            sink.AddTopLevelDeclaration(new Bpl.Axiom(dt.tok, ax, string.Format("Datatype extensional equality definition: {0}", ctor.FullName)));
-          }
-        }
-
-        // Add extensionality axiom: forall a, b :: { Dt#Equal(a, b) } Dt#Equal(a, b) <==> a == b
-        {
-          Bpl.Expr a; var aVar = BplBoundVar("a", predef.DatatypeType, out a);
-          Bpl.Expr b; var bVar = BplBoundVar("b", predef.DatatypeType, out b);
-
-          var lhs = FunctionCall(dt.tok, dtEqualName, Bpl.Type.Bool, a, b);
-          var rhs = Bpl.Expr.Eq(a, b);
-
-          var ax = BplForall(new List<Variable> { aVar, bVar }, BplTrigger(lhs), Bpl.Expr.Iff(lhs, rhs));
-          sink.AddTopLevelDeclaration(new Bpl.Axiom(dt.tok, ax, string.Format("Datatype extensionality axiom: {0}", dt.FullName)));
-        }
-      }
-
-      if (dt is CoDatatypeDecl) {
-        var codecl = (CoDatatypeDecl)dt;
-
-        Func<Bpl.Expr, Bpl.Expr> MinusOne = k => {
-          if (k == null) {
-            return null;
-          } else if (k.Type.IsInt) {
-            return Bpl.Expr.Sub(k, Bpl.Expr.Literal(1));
-          } else {
-            return FunctionCall(k.tok, "ORD#Minus", k.Type, k, FunctionCall(k.tok, "ORD#FromNat", k.Type, Bpl.Expr.Literal(1)));
-          };
-        };
-
-        Action<Bpl.Type, Action<Tuple<List<Type>, List<Type>>, List<Bpl.Variable>, List<Bpl.Expr>, List<Bpl.Expr>, Bpl.Variable, Bpl.Expr, Bpl.Expr, Bpl.Expr, Bpl.Expr, Bpl.Expr, Bpl.Expr, Bpl.Expr, Bpl.Expr>> CoAxHelper = (typeOfK, K) => {
-          Func<string, List<TypeParameter>> renew = s =>
-            Map(codecl.TypeArgs, tp =>
-              new TypeParameter(tp.tok, tp.Name + "#" + s, tp.PositionalIndex, tp.Parent));
-          List<TypeParameter> typaramsL = renew("l"), typaramsR = renew("r");
-          List<Bpl.Expr> lexprs; var lvars = MkTyParamBinders(typaramsL, out lexprs);
-          List<Bpl.Expr> rexprs; var rvars = MkTyParamBinders(typaramsR, out rexprs);
-          Func<List<TypeParameter>, List<Type>> Types = l => Map(l, tp => (Type)new UserDefinedType(tp));
-          var tyargs = Tuple.Create(Types(typaramsL), Types(typaramsR));
-
-          var vars = Concat(lvars, rvars);
-
-          Bpl.Expr k, kIsValid, kIsNonZero, kHasSuccessor, kIsLimit;
-          Bpl.Variable kVar;
-          if (typeOfK != null) {
-            kVar = BplBoundVar("k", typeOfK, out k); vars.Add(kVar);
-            if (typeOfK.IsInt) {
-              kIsValid = Bpl.Expr.Le(Bpl.Expr.Literal(0), k);
-              kIsNonZero = Bpl.Expr.Neq(Bpl.Expr.Literal(0), k);
-              kHasSuccessor = Bpl.Expr.Lt(Bpl.Expr.Literal(0), k);
-              kIsLimit = Bpl.Expr.False;
-            } else {
-              kIsValid = Bpl.Expr.True;
-              kIsNonZero = Bpl.Expr.Neq(k, FunctionCall(k.tok, "ORD#FromNat", Bpl.Type.Int, Bpl.Expr.Literal(0)));
-              kHasSuccessor = Bpl.Expr.Lt(Bpl.Expr.Literal(0), FunctionCall(k.tok, "ORD#Offset", Bpl.Type.Int, k));
-              kIsLimit = FunctionCall(k.tok, "ORD#IsLimit", Bpl.Type.Bool, k);
-            }
-          } else {
-            kVar = null; k = null;
-            kIsValid = Bpl.Expr.True;
-            kIsNonZero = Bpl.Expr.True;
-            kHasSuccessor = Bpl.Expr.True;
-            kIsLimit = Bpl.Expr.True;
-          }
-          var ly = BplBoundVar("ly", predef.LayerType, vars);
-          var d0 = BplBoundVar("d0", predef.DatatypeType, vars);
-          var d1 = BplBoundVar("d1", predef.DatatypeType, vars);
-
-          K(tyargs, vars, lexprs, rexprs, kVar, k, kIsValid, kIsNonZero, kHasSuccessor, kIsLimit, ly, d0, d1);
-        };
-
-        Action<Bpl.Type> AddAxioms = typeOfK => {
-          {
-            // Add two copies of the type parameter lists!
-            var args = MkTyParamFormals(Concat(GetTypeParams(dt), GetTypeParams(dt)), false);
-            if (typeOfK != null) {
-              args.Add(BplFormalVar(null, typeOfK, true));
-            }
-            args.Add(BplFormalVar(null, predef.LayerType, true));
-            args.Add(BplFormalVar(null, predef.DatatypeType, true));
-            args.Add(BplFormalVar(null, predef.DatatypeType, true));
-            var r = BplFormalVar(null, Bpl.Type.Bool, false);
-            var fn_nm = typeOfK != null ? CoPrefixName(codecl) : CoEqualName(codecl);
-            var fn = new Bpl.Function(dt.tok, fn_nm, args, r);
-            if (InsertChecksums) {
-              InsertChecksum(dt, fn);
-            }
-            sink.AddTopLevelDeclaration(fn);
-          }
-
-          // axiom (forall G0,...,Gn : Ty, k: int, ly : Layer, d0, d1: DatatypeType ::
-          //  { Eq(G0, .., Gn, S(ly), k, d0, d1) }
-          //  Is(d0, T(G0, .., Gn)) && Is(d1, T(G0, ... Gn)) ==>
-          //  (Eq(G0, .., Gn, S(ly), k, d0, d1)
-          //    <==>
-          //      (0 < k.Offset ==>
-          //        (d0.Nil? && d1.Nil?) ||
-          //        (d0.Cons? && d1.Cons? && d0.head == d1.head && Eq(G0, .., Gn, ly, k-1, d0.tail, d1.tail))) &&
-          //      (k != 0 && k.IsLimit ==>                        // for prefix equality only
-          //        FullEq(G0, .., Gn, ly, d0.tail, d1.tail)))    // for prefix equality only
-          CoAxHelper(typeOfK, (tyargs, vars, lexprs, rexprs, kVar, k, kIsValid, kIsNonZero, kHasSuccessor, kIsLimit, ly, d0, d1) => {
-            var eqDt = CoEqualCall(codecl, lexprs, rexprs, k, LayerSucc(ly), d0, d1);
-            var iss = BplAnd(MkIs(d0, ClassTyCon(dt, lexprs)), MkIs(d1, ClassTyCon(dt, rexprs)));
-            var body = BplImp(
-              iss,
-              BplIff(eqDt,
-                BplAnd(
-                  BplImp(kHasSuccessor, BplOr(CoPrefixEquality(dt.tok, codecl, tyargs.Item1, tyargs.Item2, MinusOne(k), ly, d0, d1))),
-                  k == null ? Bpl.Expr.True : BplImp(BplAnd(kIsNonZero, kIsLimit), CoEqualCall(codecl, tyargs.Item1, tyargs.Item2, null, ly, d0, d1)))));
-            var ax = BplForall(vars, BplTrigger(eqDt), body);
-            sink.AddTopLevelDeclaration(new Bpl.Axiom(dt.tok, ax, "Layered co-equality axiom"));
-          });
-
-          // axiom (forall G0,...,Gn : Ty, k: int, ly : Layer, d0, d1: DatatypeType ::
-          //  { Eq(G0, .., Gn, S(ly), k, d0, d1) }
-          //    0 < k ==>
-          //      (Eq(G0, .., Gn, S(ly), k, d0, d1) <==>
-          //       Eq(G0, .., Gn, ly, k, d0, d))
-          CoAxHelper(typeOfK, (tyargs, vars, lexprs, rexprs, kVar, k, kIsValid, kIsNonZero, kHasSuccessor, kIsLimit, ly, d0, d1) => {
-            var eqDtSL = CoEqualCall(codecl, lexprs, rexprs, k, LayerSucc(ly), d0, d1);
-            var eqDtL = CoEqualCall(codecl, lexprs, rexprs, k, ly, d0, d1);
-            var body = BplImp(kIsNonZero, BplIff(eqDtSL, eqDtL));
-            var ax = BplForall(vars, BplTrigger(eqDtSL), body);
-            sink.AddTopLevelDeclaration(new Bpl.Axiom(dt.tok, ax, "Unbump layer co-equality axiom"));
-          });
-        };
-
-        AddAxioms(null); // Add the above axioms for $Equal
-
-        // axiom (forall d0, d1: DatatypeType, k: int :: { $Equal(d0, d1) } :: Equal(d0, d1) <==> d0 == d1);
-        CoAxHelper(null, (tyargs, vars, lexprs, rexprs, kVar, k, kIsValid, kIsNonZero, kHasSuccessor, kIsLimit, ly, d0, d1) => {
-          var Eq = CoEqualCall(codecl, lexprs, rexprs, k, LayerSucc(ly), d0, d1);
-          var equal = Bpl.Expr.Eq(d0, d1);
-          sink.AddTopLevelDeclaration(new Axiom(dt.tok,
-            BplForall(vars, BplTrigger(Eq), BplIff(Eq, equal)),
-            "Equality for codatatypes"));
-        });
-
-        Bpl.Type theTypeOfK = predef.BigOrdinalType;
-        AddAxioms(predef.BigOrdinalType); // Add the above axioms for $PrefixEqual
-
-        // The connection between the full codatatype equality and its prefix version
-        // axiom (forall d0, d1: DatatypeType :: $Eq#Dt(d0, d1) <==>
-        //                                       (forall k: int :: 0 <= k ==> $PrefixEqual#Dt(k, d0, d1)));
-        CoAxHelper(theTypeOfK, (tyargs, vars, lexprs, rexprs, kVar, k, kIsValid, kIsNonZero, kHasSuccessor, kIsLimit, ly, d0, d1) => {
-          var Eq = CoEqualCall(codecl, lexprs, rexprs, null, LayerSucc(ly), d0, d1);
-          var PEq = CoEqualCall(codecl, lexprs, rexprs, k, LayerSucc(ly), d0, d1);
-          vars.Remove(kVar);
-          sink.AddTopLevelDeclaration(new Axiom(dt.tok,
-            BplForall(vars, BplTrigger(Eq), BplIff(Eq, BplForall(kVar, BplTrigger(PEq), BplImp(kIsValid, PEq)))),
-            "Coequality and prefix equality connection"));
-        });
-        // In addition, the following special case holds for $Eq#Dt:
-        // axiom (forall d0, d1: DatatypeType :: $Eq#Dt(d0, d1) <==
-        //                                       (forall k: int :: 0 <= k ==> $PrefixEqual#Dt(ORD#FromNat(k), d0, d1)));
-        if (!theTypeOfK.IsInt) {
-          CoAxHelper(Bpl.Type.Int, (tyargs, vars, lexprs, rexprs, kVar, k, kIsValid, kIsNonZero, kHasSuccessor, kIsLimit, ly, d0, d1) => {
-            var Eq = CoEqualCall(codecl, lexprs, rexprs, null, LayerSucc(ly), d0, d1);
-            var PEq = CoEqualCall(codecl, lexprs, rexprs, FunctionCall(k.tok, "ORD#FromNat", predef.BigOrdinalType, k), LayerSucc(ly), d0, d1);
-            vars.Remove(kVar);
-            sink.AddTopLevelDeclaration(new Axiom(dt.tok,
-              BplForall(vars, BplTrigger(Eq), BplImp(BplForall(kVar, BplTrigger(PEq), BplImp(kIsValid, PEq)), Eq)),
-              "Coequality and prefix equality connection"));
-          });
-        }
-
-        // A consequence of the definition of prefix equalities is the following:
-        // axiom (forall k, m: int, d0, d1: DatatypeType :: 0 <= k <= m && $PrefixEq#Dt(m, d0, d1) ==> $PrefixEq#0#Dt(k, d0, d1));
-        CoAxHelper(theTypeOfK, (tyargs, vars, lexprs, rexprs, kVar, k, kIsValid, kIsNonZero, kHasSuccessor, kIsLimit, ly, d0, d1) => {
-          var m = BplBoundVar("m", k.Type, vars);
-          var PEqK = CoEqualCall(codecl, lexprs, rexprs, k, LayerSucc(ly), d0, d1);
-          var PEqM = CoEqualCall(codecl, lexprs, rexprs, m, LayerSucc(ly), d0, d1);
-          Bpl.Expr kLtM;
-          if (k.Type.IsInt) {
-            kLtM = Bpl.Expr.Lt(k, m);
-          } else {
-            kLtM = FunctionCall(dt.tok, "ORD#Less", Bpl.Type.Bool, k, m);
-          }
-          sink.AddTopLevelDeclaration(new Axiom(dt.tok,
-            BplForall(vars,
-            new Bpl.Trigger(dt.tok, true, new List<Bpl.Expr> { PEqK, PEqM }),
-            BplImp(BplAnd(BplAnd(kIsValid, kLtM), PEqM), PEqK)),
-            "Prefix equality consequence"));
-        });
-
-        // With the axioms above, going from d0==d1 to a prefix equality requires going via the full codatatype
-        // equality, which in turn requires the full codatatype equality to be present.  The following axiom
-        // provides a shortcut:
-        // axiom (forall d0, d1: DatatypeType, k: int :: d0 == d1 && 0 <= k ==> $PrefixEqual#_module.Stream(k, d0, d1));
-        CoAxHelper(theTypeOfK, (tyargs, vars, lexprs, rexprs, kVar, k, kIsValid, kIsNonZero, kHasSuccessor, kIsLimit, ly, d0, d1) => {
-          var equal = Bpl.Expr.Eq(d0, d1);
-          var PEq = CoEqualCall(codecl, lexprs, rexprs, k, LayerSucc(ly), d0, d1);
-          var trigger = BplTrigger(PEq);
-          sink.AddTopLevelDeclaration(new Axiom(dt.tok,
-            BplForall(vars, trigger, BplImp(BplAnd(equal, kIsValid), PEq)), "Prefix equality shortcut"));
-        });
-      }
-    }
 
     /// <summary>
     /// Return a sequence of expressions whose conjunction denotes a memberwise equality of "dt".  Recursive
@@ -2106,201 +1490,6 @@ namespace Microsoft.Dafny {
       return fieldName;
     }
 
-    void AddClassMembers(TopLevelDeclWithMembers c, bool includeAllMethods) {
-      Contract.Requires(sink != null && predef != null);
-      Contract.Requires(c != null);
-      Contract.Ensures(fuelContext == Contract.OldValue(fuelContext));
-      Contract.Assert(VisibleInScope(c));
-
-      sink.AddTopLevelDeclaration(GetClass(c));
-      if (c is ArrayClassDecl) {
-        // classes.Add(c, predef.ClassDotArray);
-        AddAllocationAxiom(null, (ArrayClassDecl)c, true);
-      }
-
-      // Add $Is and $IsAlloc for this class :
-      //    axiom (forall p: ref, G: Ty ::
-      //       { $Is(p, TClassA(G), h) }
-      //       $Is(p, TClassA(G), h) <=> (p == null || dtype(p) == TClassA(G));
-      //    axiom (forall p: ref, h: Heap, G: Ty ::
-      //       { $IsAlloc(p, TClassA(G), h) }
-      //       $IsAlloc(p, TClassA(G), h) => (p == null || h[p, alloc]);
-      MapM(c is ClassDecl ? Bools : new List<bool>(), is_alloc => {
-        List<Bpl.Expr> tyexprs;
-        var vars = MkTyParamBinders(GetTypeParams(c), out tyexprs);
-
-        var o = BplBoundVar("$o", predef.RefType, vars);
-
-        Bpl.Expr body, is_o;
-        Bpl.Expr o_null = Bpl.Expr.Eq(o, predef.Null);
-        Bpl.Expr o_ty = ClassTyCon(c, tyexprs);
-        string name;
-
-        if (is_alloc) {
-          name = c + ": Class $IsAlloc";
-          var h = BplBoundVar("$h", predef.HeapType, vars);
-          // $IsAlloc(o, ..)
-          is_o = MkIsAlloc(o, o_ty, h);
-          body = BplIff(is_o, BplOr(o_null, IsAlloced(c.tok, h, o)));
-        } else {
-          name = c + ": Class $Is";
-          // $Is(o, ..)
-          is_o = MkIs(o, o_ty);
-          Bpl.Expr rhs;
-          if (c == program.BuiltIns.ObjectDecl) {
-            rhs = Bpl.Expr.True;
-          } else if (c is TraitDecl) {
-            //generating $o == null || implements$J(dtype(x), typeArgs)
-            var t = (TraitDecl)c;
-            var dtypeFunc = FunctionCall(o.tok, BuiltinFunction.DynamicType, null, o);
-            var implementsJ_Arguments = new List<Expr> { dtypeFunc }; // TODO: also needs type parameters
-            implementsJ_Arguments.AddRange(tyexprs);
-            Bpl.Expr implementsFunc = FunctionCall(t.tok, "implements$" + t.FullSanitizedName, Bpl.Type.Bool, implementsJ_Arguments);
-            rhs = BplOr(o_null, implementsFunc);
-          } else {
-            rhs = BplOr(o_null, DType(o, o_ty));
-          }
-          body = BplIff(is_o, rhs);
-        }
-
-        sink.AddTopLevelDeclaration(new Bpl.Axiom(c.tok, BplForall(vars, BplTrigger(is_o), body), name));
-      });
-
-      if (c is TraitDecl) {
-        //this adds: function implements$J(Ty, typeArgs): bool;
-        var arg_ref = new Bpl.Formal(c.tok, new Bpl.TypedIdent(c.tok, "ty", predef.Ty), true);
-        var vars = new List<Bpl.Variable> { arg_ref };
-        vars.AddRange(MkTyParamFormals(GetTypeParams(c)));
-        var res = new Bpl.Formal(c.tok, new Bpl.TypedIdent(c.tok, Bpl.TypedIdent.NoName, Bpl.Type.Bool), false);
-        var implement_intr = new Bpl.Function(c.tok, "implements$" + c.FullSanitizedName, vars, res);
-        sink.AddTopLevelDeclaration(implement_intr);
-      } else if (c is ClassDecl) {
-        //this adds: axiom implements$J(class.C, typeInstantiations);
-        var vars = MkTyParamBinders(GetTypeParams(c), out var tyexprs);
-
-        foreach (var parent in ((ClassDecl)c).ParentTraits) {
-          var trait = (TraitDecl)((NonNullTypeDecl)((UserDefinedType)parent).ResolvedClass).ViewAsClass;
-          var arg = ClassTyCon(c, tyexprs);
-          var args = new List<Bpl.Expr> { arg };
-          foreach (var targ in parent.TypeArgs) {
-            args.Add(TypeToTy(targ));
-          }
-          var expr = FunctionCall(c.tok, "implements$" + trait.FullSanitizedName, Bpl.Type.Bool, args);
-          var implements_axiom = new Bpl.Axiom(c.tok, BplForall(vars, null, expr));
-          sink.AddTopLevelDeclaration(implements_axiom);
-        }
-      }
-
-      foreach (MemberDecl member in c.Members.FindAll(VisibleInScope)) {
-        Contract.Assert(isAllocContext == null);
-        currentDeclaration = member;
-        if (member is Field) {
-          Field f = (Field)member;
-          if (f is ConstantField) {
-            // The following call has the side effect of idempotently creating and adding the function to the sink's top-level declarations
-            Contract.Assert(currentModule == null);
-            currentModule = f.EnclosingClass.EnclosingModuleDefinition;
-            var oldFuelContext = fuelContext;
-            fuelContext = FuelSetting.NewFuelContext(f);
-            var boogieFunction = GetReadonlyField(f);
-            fuelContext = oldFuelContext;
-            currentModule = null;
-            AddAllocationAxiom(f, c);
-          } else {
-            if (f.IsMutable) {
-              Bpl.Constant fc = GetField(f);
-              sink.AddTopLevelDeclaration(fc);
-            } else {
-              Bpl.Function ff = GetReadonlyField(f);
-              if (ff != predef.ArrayLength) {
-                sink.AddTopLevelDeclaration(ff);
-              }
-            }
-            AddAllocationAxiom(f, c);
-          }
-
-        } else if (member is Function function) {
-          AddFunction_Top(function, includeAllMethods);
-        } else if (member is Method method) {
-          AddMethod_Top(method, false, includeAllMethods);
-        } else {
-          Contract.Assert(false); throw new cce.UnreachableException();  // unexpected member
-        }
-      }
-    }
-
-    void AddFunction_Top(Function f, bool includeAllMethods) {
-      FuelContext oldFuelContext = this.fuelContext;
-      this.fuelContext = FuelSetting.NewFuelContext(f);
-      isAllocContext = new IsAllocContext(true);
-
-      AddClassMember_Function(f);
-
-      if (!f.IsBuiltin && InVerificationScope(f)) {
-        AddWellformednessCheck(f);
-        if (f.OverriddenFunction != null) { //it means that f is overriding its associated parent function
-          AddFunctionOverrideCheckImpl(f);
-        }
-      }
-      if (f is ExtremePredicate cop) {
-        AddClassMember_Function(cop.PrefixPredicate);
-        // skip the well-formedness check, because it has already been done for the extreme predicate
-      } else if (f.ByMethodDecl != null) {
-        AddMethod_Top(f.ByMethodDecl, true, includeAllMethods);
-      }
-
-      this.fuelContext = oldFuelContext;
-      isAllocContext = null;
-    }
-
-    void AddMethod_Top(Method m, bool isByMethod, bool includeAllMethods) {
-      if (!includeAllMethods && !InVerificationScope(m) && !referencedMembers.Contains(m)) {
-        // do nothing
-        return;
-      }
-
-      FuelContext oldFuelContext = this.fuelContext;
-      this.fuelContext = FuelSetting.NewFuelContext(m);
-
-      // wellformedness check for method specification
-      if (m.EnclosingClass is IteratorDecl && m == ((IteratorDecl)m.EnclosingClass).Member_MoveNext) {
-        // skip the well-formedness check, because it has already been done for the iterator
-      } else {
-        if (!isByMethod) {
-          var proc = AddMethod(m, MethodTranslationKind.SpecWellformedness);
-          sink.AddTopLevelDeclaration(proc);
-          if (InVerificationScope(m)) {
-            AddMethodImpl(m, proc, true);
-          }
-        }
-        if (m.OverriddenMethod != null && InVerificationScope(m)) //method has overrided a parent method
-        {
-          var procOverrideChk = AddMethod(m, MethodTranslationKind.OverrideCheck);
-          sink.AddTopLevelDeclaration(procOverrideChk);
-          AddMethodOverrideCheckImpl(m, procOverrideChk);
-        }
-      }
-      // the method spec itself
-      if (!isByMethod) {
-        sink.AddTopLevelDeclaration(AddMethod(m, MethodTranslationKind.Call));
-      }
-      if (m is ExtremeLemma) {
-        // Let the CoCall and Impl forms to use m.PrefixLemma signature and specification (and
-        // note that m.PrefixLemma.Body == m.Body.
-        m = ((ExtremeLemma)m).PrefixLemma;
-        sink.AddTopLevelDeclaration(AddMethod(m, MethodTranslationKind.CoCall));
-      }
-      if (m.Body != null && InVerificationScope(m)) {
-        // ...and its implementation
-        assertionCount = 0;
-        var proc = AddMethod(m, MethodTranslationKind.Implementation);
-        sink.AddTopLevelDeclaration(proc);
-        AddMethodImpl(m, proc, false);
-      }
-      Reset();
-      this.fuelContext = oldFuelContext;
-    }
-
     /// <summary>
     /// Returns true if the body of function "f" is available in module "context".
     /// This happens when the following conditions all hold:
@@ -2319,47 +1508,6 @@ namespace Microsoft.Dafny {
     static bool IsOpaqueRevealLemma(Method m) {
       Contract.Requires(m != null);
       return Attributes.Contains(m.Attributes, "opaque_reveal");
-    }
-
-    private void AddClassMember_Function(Function f) {
-      Contract.Ensures(currentModule == null && codeContext == null);
-      Contract.Ensures(currentModule == null && codeContext == null);
-
-      currentModule = f.EnclosingClass.EnclosingModuleDefinition;
-      codeContext = f;
-
-      // declare function
-      AddFunction(f);
-      // add synonym axiom
-      if (f.IsFuelAware()) {
-        AddLayerSynonymAxiom(f);
-        AddFuelSynonymAxiom(f);
-      }
-      // add frame axiom
-      if (AlwaysUseHeap || f.ReadsHeap) {
-        AddFrameAxiom(f);
-      }
-      // add consequence axiom
-      AddFunctionConsequenceAxiom(f, f.Ens);
-      // add definition axioms, suitably specialized for literals
-      if (f.Body != null && RevealedInScope(f)) {
-        AddFunctionAxiom(f, f.Body.Resolved);
-      } else {
-        // for body-less functions, at least generate its #requires function
-        var b = FunctionAxiom(f, null, null);
-        Contract.Assert(b == null);
-      }
-      // for a function in a class C that overrides a function in a trait J, add an axiom that connects J.F and C.F
-      if (f.OverriddenFunction != null) {
-        sink.AddTopLevelDeclaration(FunctionOverrideAxiom(f.OverriddenFunction, f));
-      }
-
-      // supply the connection between least/greatest predicates and prefix predicates
-      if (f is ExtremePredicate) {
-        AddPrefixPredicateAxioms(((ExtremePredicate)f).PrefixPredicate);
-      }
-
-      Reset();
     }
 
     void AddIteratorSpecAndBody(IteratorDecl iter) {
@@ -2763,12 +1911,12 @@ namespace Microsoft.Dafny {
       }
     }
 
-    void AddFunctionAxiom(Function f, Expression body) {
+    void AddFunctionAxiom(Bpl.Function boogieFunction, Function f, Expression body) {
       Contract.Requires(f != null);
       Contract.Requires(body != null);
 
-      var ax = FunctionAxiom(f, body, null);
-      sink.AddTopLevelDeclaration(ax);
+      var ax = GetFunctionAxiom(f, body, null);
+      AddOtherDefinition(boogieFunction, ax);
       // TODO(namin) Is checking f.Reads.Count==0 excluding Valid() of BinaryTree in the right way?
       //             I don't see how this in the decreasing clause would help there.
       if (!(f is ExtremePredicate) && f.CoClusterTarget == Function.CoCallClusterInvolvement.None && f.Reads.Count == 0) {
@@ -2800,16 +1948,16 @@ namespace Microsoft.Dafny {
 
         Contract.Assert(decs.Count <= allFormals.Count);
         if (0 < decs.Count && decs.Count < allFormals.Count) {
-          ax = FunctionAxiom(f, body, decs);
-          sink.AddTopLevelDeclaration(ax);
+          var decreasesAxiom = GetFunctionAxiom(f, body, decs);
+          AddOtherDefinition(boogieFunction, decreasesAxiom);
         }
 
-        ax = FunctionAxiom(f, body, allFormals);
-        sink.AddTopLevelDeclaration(ax);
+        var formalsAxiom = GetFunctionAxiom(f, body, allFormals);
+        AddOtherDefinition(boogieFunction, formalsAxiom);
       }
     }
 
-    void AddFunctionConsequenceAxiom(Function f, List<AttributedExpression> ens) {
+    void AddFunctionConsequenceAxiom(Boogie.Function boogieFunction, Function f, List<AttributedExpression> ens) {
       Contract.Requires(f != null);
       Contract.Requires(predef != null);
       Contract.Requires(f.EnclosingClass != null);
@@ -2980,7 +2128,8 @@ namespace Microsoft.Dafny {
       Bpl.Expr ax = BplForall(f.tok, new List<Bpl.TypeVariable>(), formals, null, tr, Bpl.Expr.Imp(ante, post));
       var activate = AxiomActivation(f, etran);
       string comment = "consequence axiom for " + f.FullSanitizedName;
-      sink.AddTopLevelDeclaration(new Bpl.Axiom(f.tok, Bpl.Expr.Imp(activate, ax), comment));
+      var consequenceAxiom = new Bpl.Axiom(f.tok, Bpl.Expr.Imp(activate, ax), comment);
+      AddOtherDefinition(boogieFunction, consequenceAxiom);
 
       if (CommonHeapUse && !readsHeap) {
         whr = GetWhereClause(f.tok, funcAppl, f.ResultType, etranHeap, NOALLOC, true);
@@ -2990,7 +2139,8 @@ namespace Microsoft.Dafny {
           ante = BplAnd(ante, goodHeap);
           ax = BplForall(f.tok, new List<Bpl.TypeVariable>(), formals, null, BplTrigger(whr), Bpl.Expr.Imp(ante, whr));
           activate = AxiomActivation(f, etran);
-          sink.AddTopLevelDeclaration(new Bpl.Axiom(f.tok, Bpl.Expr.Imp(activate, ax)));
+          var heapConsequenceAxiom = new Bpl.Axiom(f.tok, Bpl.Expr.Imp(activate, ax));
+          AddOtherDefinition(boogieFunction, heapConsequenceAxiom);
         }
       }
     }
@@ -3013,11 +2163,11 @@ namespace Microsoft.Dafny {
     /// The list of formals "lits" is allowed to contain an object of type ThisSurrogate, which indicates that
     /// the receiver parameter of the function should be included among the lit formals.
     /// </summary>
-    Bpl.Axiom FunctionAxiom(Function f, Expression/*?*/ body, List<Formal>/*?*/ lits) {
+    private Axiom GetFunctionAxiom(Function f, Expression body, List<Formal> lits) {
       Contract.Requires(f != null);
       Contract.Requires(predef != null);
       Contract.Requires(f.EnclosingClass != null);
-      Contract.Ensures((Contract.Result<Bpl.Axiom>() == null) == (body == null));  // return null iff body is null
+      Contract.Ensures((Contract.Result<Axiom>() == null) == (body == null)); // return null iff body is null
 
       // This method generates the Definition Axiom, suitably modified according to the optional "lits".
       //
@@ -3075,6 +2225,7 @@ namespace Microsoft.Dafny {
       foreach (AttributedExpression e in f.Req) {
         readsHeap = readsHeap || UsesHeap(e.E);
       }
+
       if (body != null && UsesHeap(body)) {
         readsHeap = true;
       }
@@ -3087,9 +2238,9 @@ namespace Microsoft.Dafny {
           f.ReadsHeap ? new Bpl.IdentifierExpr(f.tok, predef.HeapVarName, predef.HeapType) : null,
           new Bpl.IdentifierExpr(f.tok, bvPrevHeap));
       } else {
-        etran = readsHeap ?
-          new ExpressionTranslator(this, predef, f.tok) :
-          new ExpressionTranslator(this, predef, (Bpl.Expr)null);
+        etran = readsHeap
+          ? new ExpressionTranslator(this, predef, f.tok)
+          : new ExpressionTranslator(this, predef, (Bpl.Expr)null);
       }
 
       // quantify over the type arguments, and add them first to the arguments
@@ -3121,21 +2272,25 @@ namespace Microsoft.Dafny {
         // ante:  $IsGoodHeap($prevHeap) &&
         ante = BplAnd(ante, FunctionCall(f.tok, BuiltinFunction.IsGoodHeap, null, etran.Old.HeapExpr));
       }
+
       Bpl.Expr goodHeap = null;
       var bv = new Bpl.BoundVariable(f.tok, new Bpl.TypedIdent(f.tok, predef.HeapVarName, predef.HeapType));
       if (AlwaysUseHeap || f.ReadsHeap) {
         funcFormals.Add(bv);
       }
+
       if (AlwaysUseHeap || f.ReadsHeap) {
         args.Add(new Bpl.IdentifierExpr(f.tok, bv));
         reqFuncArguments.Add(new Bpl.IdentifierExpr(f.tok, bv));
       }
+
       // ante:  $IsGoodHeap($Heap) && $HeapSucc($prevHeap, $Heap) && this != null && formals-have-the-expected-types &&
       if (readsHeap) {
         forallFormals.Add(bv);
         goodHeap = FunctionCall(f.tok, BuiltinFunction.IsGoodHeap, null, etran.HeapExpr);
         ante = BplAnd(ante, goodHeap);
       }
+
       if (f is TwoStateFunction && f.ReadsHeap) {
         ante = BplAnd(ante, HeapSucc(etran.Old.HeapExpr, etran.HeapExpr));
       }
@@ -3157,6 +2312,7 @@ namespace Microsoft.Dafny {
         } else {
           args.Add(bvThisIdExpr);
         }
+
         // add well-typedness conjunct to antecedent
         Type thisType = Resolver.GetReceiverType(f.tok, f);
         Bpl.Expr wh = Bpl.Expr.And(
@@ -3166,11 +2322,12 @@ namespace Microsoft.Dafny {
       }
 
       var typeMap = new Dictionary<TypeParameter, Type>();
-      var anteReqAxiom = ante;  // note that antecedent so far is the same for #requires axioms, even the receiver parameter of a two-state function
+      var anteReqAxiom = ante; // note that antecedent so far is the same for #requires axioms, even the receiver parameter of a two-state function
       var substMap = new Dictionary<IVariable, Expression>();
       foreach (Formal p in f.Formals) {
         var pType = Resolver.SubstType(p.Type, typeMap);
-        bv = new Bpl.BoundVariable(p.tok, new Bpl.TypedIdent(p.tok, p.AssignUniqueName(currentDeclaration.IdGenerator), TrType(pType)));
+        bv = new Bpl.BoundVariable(p.tok,
+          new Bpl.TypedIdent(p.tok, p.AssignUniqueName(currentDeclaration.IdGenerator), TrType(pType)));
         forallFormals.Add(bv);
         funcFormals.Add(bv);
         reqFuncArguments.Add(new Bpl.IdentifierExpr(f.tok, bv));
@@ -3178,24 +2335,32 @@ namespace Microsoft.Dafny {
         if (lits != null && lits.Contains(p) && !substMap.ContainsKey(p)) {
           args.Add(Lit(formal));
           var ie = new IdentifierExpr(p.tok, p.AssignUniqueName(f.IdGenerator));
-          ie.Var = p; ie.Type = ie.Var.Type;
+          ie.Var = p;
+          ie.Type = ie.Var.Type;
           var l = new UnaryOpExpr(p.tok, UnaryOpExpr.Opcode.Lit, ie);
           l.Type = ie.Var.Type;
           substMap.Add(p, l);
         } else {
           args.Add(formal);
         }
+
         // add well-typedness conjunct to antecedent
         Bpl.Expr wh = GetWhereClause(p.tok, formal, pType, p.IsOld ? etran.Old : etran, NOALLOC);
-        if (wh != null) { ante = BplAnd(ante, wh); }
+        if (wh != null) {
+          ante = BplAnd(ante, wh);
+        }
+
         wh = GetWhereClause(p.tok, formal, pType, etran, NOALLOC);
-        if (wh != null) { anteReqAxiom = BplAnd(anteReqAxiom, wh); }
+        if (wh != null) {
+          anteReqAxiom = BplAnd(anteReqAxiom, wh);
+        }
       }
 
       Bpl.Expr pre = Bpl.Expr.True;
       foreach (AttributedExpression req in f.Req) {
         pre = BplAnd(pre, etran.TrExpr(Substitute(req.E, receiverReplacement, substMap)));
       }
+
       var preReqAxiom = pre;
       if (f is TwoStateFunction) {
         // Checked preconditions that old parameters really existed in previous state
@@ -3206,8 +2371,10 @@ namespace Microsoft.Dafny {
             var dafnyFormalIdExpr = new IdentifierExpr(formal.tok, formal);
             preRA = BplAnd(preRA, MkIsAlloc(etran.TrExpr(dafnyFormalIdExpr), formal.Type, etran.Old.HeapExpr));
           }
+
           index++;
         }
+
         preReqAxiom = BplAnd(preRA, pre);
       }
 
@@ -3222,18 +2389,21 @@ namespace Microsoft.Dafny {
         var appl = FunctionCall(f.tok, RequiresName(f), Bpl.Type.Bool, reqFuncArguments);
         Bpl.Trigger trig = BplTriggerHeap(this, f.tok, appl, readsHeap ? etran.HeapExpr : null);
         // axiom (forall params :: { f#requires(params) }  ante ==> f#requires(params) == pre);
-        sink.AddTopLevelDeclaration(new Axiom(f.tok,
+        AddIncludeDepAxiom(new Axiom(f.tok,
           BplForall(forallFormals, trig, BplImp(anteReqAxiom, Bpl.Expr.Eq(appl, preReqAxiom))),
           "#requires axiom for " + f.FullSanitizedName));
       }
+
       if (body == null || !RevealedInScope(f)) {
         return null;
       }
 
       // useViaContext: (mh != ModuleContextHeight || fh != FunctionContextHeight)
       ModuleDefinition mod = f.EnclosingClass.EnclosingModuleDefinition;
-      Bpl.Expr useViaContext = !InVerificationScope(f) ? (Bpl.Expr)Bpl.Expr.True :
-        Bpl.Expr.Neq(Bpl.Expr.Literal(mod.CallGraph.GetSCCRepresentativePredecessorCount(f)), etran.FunctionContextHeight());
+      Bpl.Expr useViaContext = !InVerificationScope(f)
+        ? (Bpl.Expr)Bpl.Expr.True
+        : Bpl.Expr.Neq(Bpl.Expr.Literal(mod.CallGraph.GetSCCRepresentativePredecessorCount(f)),
+          etran.FunctionContextHeight());
       // ante := (useViaContext && typeAnte && pre)
       ante = BplAnd(useViaContext, BplAnd(ante, pre));
 
@@ -3257,6 +2427,7 @@ namespace Microsoft.Dafny {
           //  funcArgs.Add(ly);
           //}
         }
+
         funcArgs.AddRange(args);
         funcAppl = new Bpl.NAryExpr(f.tok, new Bpl.FunctionCall(funcID), funcArgs);
       }
@@ -3271,23 +2442,29 @@ namespace Microsoft.Dafny {
           var pp = (PrefixPredicate)f;
           bodyWithSubst = PrefixSubstitution(pp, bodyWithSubst);
         }
+
         Bpl.Expr ly = null;
         if (layer != null) {
           ly = new Bpl.IdentifierExpr(f.tok, layer);
-          if (lits != null) {   // Lit axiom doesn't consume any fuel
+          if (lits != null) {
+            // Lit axiom doesn't consume any fuel
             ly = LayerSucc(ly);
           }
         }
+
         var etranBody = layer == null ? etran : etran.LimitedFunctions(f, ly);
         var trbody = etranBody.TrExpr(bodyWithSubst);
         tastyVegetarianOption = BplAnd(CanCallAssumption(bodyWithSubst, etranBody),
           BplAnd(TrFunctionSideEffect(bodyWithSubst, etranBody), Bpl.Expr.Eq(funcAppl, trbody)));
       }
+
       QKeyValue kv = null;
       if (lits != null) {
         kv = new QKeyValue(f.tok, "weight", new List<object>() { Bpl.Expr.Literal(3) }, null);
       }
-      Bpl.Expr ax = BplForall(f.tok, new List<Bpl.TypeVariable>(), forallFormals, kv, tr, Bpl.Expr.Imp(ante, tastyVegetarianOption));
+
+      Bpl.Expr ax = BplForall(f.tok, new List<Bpl.TypeVariable>(), forallFormals, kv, tr,
+        Bpl.Expr.Imp(ante, tastyVegetarianOption));
       var activate = AxiomActivation(f, etran);
       string comment;
       comment = "definition axiom for " + f.FullSanitizedName;
@@ -3298,157 +2475,13 @@ namespace Microsoft.Dafny {
           comment += " for decreasing-related literals";
         }
       }
+
       if (RevealedInScope(f)) {
         comment += " (revealed)";
       } else {
         comment += " (opaque)";
       }
-      return new Bpl.Axiom(f.tok, Bpl.Expr.Imp(activate, ax), comment);
-    }
-
-    /// <summary>
-    /// Essentially, the function override axiom looks like:
-    ///   axiom (forall $heap: HeapType, typeArgs: Ty, this: ref, x#0: int, fuel: LayerType ::
-    ///     { J.F(fuel, $heap, G(typeArgs), this, x#0), C.F(fuel, $heap, typeArgs, this, x#0) }
-    ///     { J.F(fuel, $heap, G(typeArgs), this, x#0), $Is(this, C) }
-    ///     this != null && $Is(this, C)
-    ///     ==>
-    ///     J.F(fuel, $heap, G(typeArgs), this, x#0) == C.F(fuel, $heap, typeArgs, this, x#0));
-    /// (without the other usual antecedents).  Essentially, the override gives a part of the body of the
-    /// trait's function, so we call FunctionAxiom to generate a conditional axiom (that is, we pass in the "overridingFunction"
-    /// parameter to FunctionAxiom, which will add 'dtype(this) == class.C' as an additional antecedent) for a
-    /// body of 'C.F(this, x#0)'.
-    /// </summary>
-    Bpl.Axiom FunctionOverrideAxiom(Function f, Function overridingFunction) {
-      Contract.Requires(f != null);
-      Contract.Requires(overridingFunction != null);
-      Contract.Requires(predef != null);
-      Contract.Requires(f.EnclosingClass != null);
-      Contract.Requires(!f.IsStatic);
-      Contract.Requires(overridingFunction.EnclosingClass is TopLevelDeclWithMembers);
-      Contract.Ensures(Contract.Result<Bpl.Axiom>() != null);
-
-      bool readsHeap = AlwaysUseHeap || f.ReadsHeap || overridingFunction.ReadsHeap;
-
-      ExpressionTranslator etran;
-      Bpl.BoundVariable bvPrevHeap = null;
-      if (f is TwoStateFunction) {
-        bvPrevHeap = new Bpl.BoundVariable(f.tok, new Bpl.TypedIdent(f.tok, "$prevHeap", predef.HeapType));
-        etran = new ExpressionTranslator(this, predef,
-          f.ReadsHeap ? new Bpl.IdentifierExpr(f.tok, predef.HeapVarName, predef.HeapType) : null,
-          new Bpl.IdentifierExpr(f.tok, bvPrevHeap));
-      } else if (readsHeap) {
-        etran = new ExpressionTranslator(this, predef, f.tok);
-      } else {
-        etran = new ExpressionTranslator(this, predef, (Bpl.Expr)null);
-      }
-
-      // "forallFormals" is built to hold the bound variables of the quantification
-      // argsJF are the arguments to J.F (the function in the trait)
-      // argsCF are the arguments to C.F (the overriding function)
-      var forallFormals = new List<Bpl.Variable>();
-      var argsJF = new List<Bpl.Expr>();
-      var argsCF = new List<Bpl.Expr>();
-
-      // Add type arguments
-      forallFormals.AddRange(MkTyParamBinders(GetTypeParams(overridingFunction), out _));
-      argsJF.AddRange(GetTypeArguments(f, overridingFunction).ConvertAll(TypeToTy));
-      argsCF.AddRange(GetTypeArguments(overridingFunction, null).ConvertAll(TypeToTy));
-
-      // Add the fuel argument
-      if (f.IsFuelAware()) {
-        Contract.Assert(overridingFunction.IsFuelAware());  // f.IsFuelAware() ==> overridingFunction.IsFuelAware()
-        var fuel = new Bpl.BoundVariable(f.tok, new Bpl.TypedIdent(f.tok, "$fuel", predef.LayerType));
-        forallFormals.Add(fuel);
-        var ly = new Bpl.IdentifierExpr(f.tok, fuel);
-        argsJF.Add(ly);
-        argsCF.Add(ly);
-      } else if (overridingFunction.IsFuelAware()) {
-        // We can't use a bound variable $fuel, because then one of the triggers won't be mentioning this $fuel.
-        // Instead, we do the next best thing: use the literal $LZ.
-        var ly = new Bpl.IdentifierExpr(f.tok, "$LZ", predef.LayerType); // $LZ
-        argsCF.Add(ly);
-      }
-
-      // Add heap arguments
-      if (f is TwoStateFunction) {
-        Contract.Assert(bvPrevHeap != null);
-        forallFormals.Add(bvPrevHeap);
-        argsJF.Add(etran.Old.HeapExpr);
-        argsCF.Add(etran.Old.HeapExpr);
-      }
-      if (AlwaysUseHeap || f.ReadsHeap || overridingFunction.ReadsHeap) {
-        var heap = new Bpl.BoundVariable(f.tok, new Bpl.TypedIdent(f.tok, predef.HeapVarName, predef.HeapType));
-        forallFormals.Add(heap);
-        if (AlwaysUseHeap || f.ReadsHeap) {
-          argsJF.Add(new Bpl.IdentifierExpr(f.tok, heap));
-        }
-        if (AlwaysUseHeap || overridingFunction.ReadsHeap) {
-          argsCF.Add(new Bpl.IdentifierExpr(overridingFunction.tok, heap));
-        }
-      }
-
-      // Add receiver parameter
-      Type thisType = Resolver.GetReceiverType(f.tok, overridingFunction);
-      var bvThis = new Bpl.BoundVariable(f.tok, new Bpl.TypedIdent(f.tok, etran.This, TrType(thisType)));
-      forallFormals.Add(bvThis);
-      var bvThisExpr = new Bpl.IdentifierExpr(f.tok, bvThis);
-      argsJF.Add(bvThisExpr);
-      argsCF.Add(bvThisExpr);
-      // $Is(this, C)
-      var isOfSubtype = GetWhereClause(overridingFunction.tok, bvThisExpr, thisType, f is TwoStateFunction ? etran.Old : etran, IsAllocType.NEVERALLOC);
-
-      // Add other arguments
-      var typeMap = GetTypeArgumentSubstitutionMap(f, overridingFunction);
-      foreach (Formal p in f.Formals) {
-        var pType = Resolver.SubstType(p.Type, typeMap);
-        var bv = new Bpl.BoundVariable(p.tok, new Bpl.TypedIdent(p.tok, p.AssignUniqueName(currentDeclaration.IdGenerator), TrType(pType)));
-        forallFormals.Add(bv);
-        var jfArg = new Bpl.IdentifierExpr(p.tok, bv);
-        argsJF.Add(ModeledAsBoxType(p.Type) ? BoxIfUnboxed(jfArg, pType) : jfArg);
-        argsCF.Add(new Bpl.IdentifierExpr(p.tok, bv));
-      }
-
-      // useViaContext: (mh != ModuleContextHeight || fh != FunctionContextHeight)
-      ModuleDefinition mod = f.EnclosingClass.EnclosingModuleDefinition;
-      Bpl.Expr useViaContext = !InVerificationScope(overridingFunction) ? (Bpl.Expr)Bpl.Expr.True :
-        Bpl.Expr.Neq(Bpl.Expr.Literal(mod.CallGraph.GetSCCRepresentativePredecessorCount(overridingFunction)), etran.FunctionContextHeight());
-
-      Bpl.Expr funcAppl;
-      {
-        var funcID = new Bpl.IdentifierExpr(f.tok, f.FullSanitizedName, TrType(f.ResultType));
-        funcAppl = new Bpl.NAryExpr(f.tok, new Bpl.FunctionCall(funcID), argsJF);
-      }
-      Bpl.Expr overridingFuncAppl;
-      {
-        var funcID = new Bpl.IdentifierExpr(overridingFunction.tok, overridingFunction.FullSanitizedName, TrType(overridingFunction.ResultType));
-        overridingFuncAppl = new Bpl.NAryExpr(overridingFunction.tok, new Bpl.FunctionCall(funcID), argsCF);
-      }
-
-      // Build the triggers
-      // { f(Succ(s), args), f'(Succ(s), args') }
-      Bpl.Trigger tr = BplTriggerHeap(this, overridingFunction.tok,
-        funcAppl,
-        readsHeap ? etran.HeapExpr : null,
-        overridingFuncAppl);
-      // { f(Succ(s), args), $Is(this, T') }
-      var exprs = new List<Bpl.Expr>() { funcAppl, isOfSubtype };
-      if (readsHeap) {
-        exprs.Add(FunctionCall(overridingFunction.tok, BuiltinFunction.IsGoodHeap, null, etran.HeapExpr));
-      }
-      tr = new Bpl.Trigger(overridingFunction.tok, true, exprs, tr);
-
-      // The equality that is what it's all about
-      var synonyms = Bpl.Expr.Eq(
-        funcAppl,
-        ModeledAsBoxType(f.ResultType) ? BoxIfUnboxed(overridingFuncAppl, overridingFunction.ResultType) : overridingFuncAppl);
-
-      // The axiom
-      Bpl.Expr ax = BplForall(f.tok, new List<Bpl.TypeVariable>(), forallFormals, null, tr,
-        Bpl.Expr.Imp(Bpl.Expr.And(ReceiverNotNull(bvThisExpr), isOfSubtype), synonyms));
-      var activate = AxiomActivation(f, etran);
-      string comment = "override axiom for " + f.FullSanitizedName + " in class " + overridingFunction.EnclosingClass.FullSanitizedName;
-      return new Bpl.Axiom(f.tok, Bpl.Expr.Imp(activate, ax), comment);
+      return new Axiom(f.tok, Bpl.Expr.Imp(activate, ax), comment);
     }
 
     Bpl.Type TrReceiverType(MemberDecl f) {
@@ -3703,7 +2736,7 @@ namespace Microsoft.Dafny {
 
       Bpl.Trigger tr = new Bpl.Trigger(f.tok, true, new List<Bpl.Expr> { funcAppl1 });
       Bpl.Expr ax = new Bpl.ForallExpr(f.tok, new List<Bpl.TypeVariable>(), formals, null, tr, Bpl.Expr.Eq(funcAppl1, funcAppl0));
-      sink.AddTopLevelDeclaration(new Bpl.Axiom(f.tok, ax, "layer synonym axiom"));
+      AddIncludeDepAxiom(new Bpl.Axiom(f.tok, ax, "layer synonym axiom"));
     }
 
     void AddFuelSynonymAxiom(Function f) {
@@ -3769,7 +2802,7 @@ namespace Microsoft.Dafny {
 
       Bpl.Trigger tr = new Bpl.Trigger(f.tok, true, new List<Bpl.Expr> { funcAppl2 });
       Bpl.Expr ax = new Bpl.ForallExpr(f.tok, new List<Bpl.TypeVariable>(), formals, null, tr, Bpl.Expr.Eq(funcAppl1, funcAppl0));
-      sink.AddTopLevelDeclaration(new Bpl.Axiom(f.tok, ax, "fuel synonym axiom"));
+      AddIncludeDepAxiom(new Bpl.Axiom(f.tok, ax, "fuel synonym axiom"));
     }
 
     /// <summary>
@@ -3945,7 +2978,7 @@ namespace Microsoft.Dafny {
 
       var trigger2 = new Bpl.Trigger(tok, true, new List<Bpl.Expr> { prefixPred_K, prefixPred_M });
       var monotonicity = new Bpl.ForallExpr(tok, moreBvs, trigger2, BplImp(smaller, direction));
-      sink.AddTopLevelDeclaration(new Bpl.Axiom(tok, Bpl.Expr.Imp(activation, monotonicity),
+      AddRootAxiom(new Bpl.Axiom(tok, Bpl.Expr.Imp(activation, monotonicity),
         "prefix predicate monotonicity axiom"));
 #endif
       // A more targeted monotonicity axiom used to increase the power of automation for proving the limit case for
@@ -3973,232 +3006,6 @@ namespace Microsoft.Dafny {
         var monotonicity = new Bpl.ForallExpr(tok, moreBvs, trigger3, BplImp(kLessM, direction));
         sink.AddTopLevelDeclaration(new Bpl.Axiom(tok, Bpl.Expr.Imp(activation, monotonicity),
           "targeted prefix predicate monotonicity axiom"));
-      }
-    }
-
-    /// <summary>
-    /// For a non-static field "f" in a class "c(G)", generate:
-    ///     // type axiom:
-    ///     // If "G" is empty, then TClassA(G) is omitted from trigger.
-    ///     // If "c" is an array declaration, then the bound variables also include the index variables "ii" and "h[o, f]" has the form "h[o, Index(ii)]".
-    ///     // If "f" is readonly, then "h[o, f]" has the form "f(o)" (for special fields) or "f(G,o)" (for programmer-declared const fields),
-    ///     // so "h" and $IsHeap(h) are omitted.
-    ///     axiom fh < FunctionContextHeight ==>
-    ///       (forall o: ref, h: Heap, G : Ty ::
-    ///         { h[o, f], TClassA(G) }  // if "f" is a const, omit TClassA(G) from the trigger and just use { f(G,o) }
-    ///         $IsHeap(h) &&
-    ///         o != null && $Is(o, TClassA(G))  // or dtype(o) = TClassA(G)
-    ///         ==>
-    ///         $Is(h[o, f], TT(PP)));
-    ///
-    ///     // allocation axiom:
-    ///     // As above for "G" and "ii", but "h" is included no matter what.
-    ///     axiom fh < FunctionContextHeight ==>
-    ///       (forall o: ref, h: Heap, G : Ty ::
-    ///         { h[o, f], TClassA(G) }  // if "f" is a const, use the trigger { f(G,o), h[o, alloc] }; for other readonly fields, use { f(o), h[o, alloc], TClassA(G) }
-    ///         $IsHeap(h) &&
-    ///         o != null && $Is(o, TClassA(G)) &&  // or dtype(o) = TClassA(G)
-    ///         h[o, alloc]
-    ///         ==>
-    ///         $IsAlloc(h[o, f], TT(PP), h));
-    ///
-    /// For a static (necessarily "const") field "f" in a class "c(G)", the expression corresponding to "h[o, f]" or "f(G,o)" above is "f(G)",
-    /// so generate:
-    ///     // type axiom:
-    ///     axiom fh < FunctionContextHeight ==>
-    ///       (forall G : Ty ::
-    ///         { f(G) }
-    ///         $Is(f(G), TT(PP)));
-    ///     // Or in the case where G is empty:
-    ///     axiom $Is(f(G), TT);
-    ///
-    ///     // allocation axiom:
-    ///     axiom fh < FunctionContextHeight ==>
-    ///       (forall h: Heap, G : Ty ::
-    ///         { $IsAlloc(f(G), TT(PP), h) }
-    ///         $IsHeap(h)
-    ///       ==>
-    ///         $IsAlloc(f(G), TT(PP), h));
-    ///
-    ///
-    /// The axioms above could be optimised to something along the lines of:
-    ///     axiom fh < FunctionContextHeight ==>
-    ///       (forall o: ref, h: Heap ::
-    ///         { h[o, f] }
-    ///         $IsHeap(h) && o != null && Tag(dtype(o)) = TagClass
-    ///         ==>
-    ///         (h[o, alloc] ==> $IsAlloc(h[o, f], TT(TClassA_Inv_i(dtype(o)),..), h)) &&
-    ///         $Is(h[o, f], TT(TClassA_Inv_i(dtype(o)),..), h);
-    /// <summary>
-    void AddAllocationAxiom(Field f, TopLevelDeclWithMembers c, bool is_array = false) {
-      Contract.Requires(c != null);
-      // IFF you're adding the array axioms, then the field should be null
-      Contract.Requires(is_array == (f == null));
-      Contract.Requires(sink != null && predef != null);
-
-      Bpl.Expr heightAntecedent = Bpl.Expr.True;
-      if (f is ConstantField) {
-        var cf = (ConstantField)f;
-        AddWellformednessCheck(cf);
-        if (InVerificationScope(cf)) {
-          var etran = new ExpressionTranslator(this, predef, f.tok);
-          heightAntecedent = Bpl.Expr.Lt(Bpl.Expr.Literal(cf.EnclosingModule.CallGraph.GetSCCRepresentativePredecessorCount(cf)), etran.FunctionContextHeight());
-        }
-      }
-
-      var bvsTypeAxiom = new List<Bpl.Variable>();
-      var bvsAllocationAxiom = new List<Bpl.Variable>();
-
-      // G
-      List<Bpl.Expr> tyexprs;
-      var tyvars = MkTyParamBinders(GetTypeParams(c), out tyexprs);
-      bvsTypeAxiom.AddRange(tyvars);
-      bvsAllocationAxiom.AddRange(tyvars);
-
-      if (f is ConstantField && f.IsStatic) {
-        var oDotF = new Bpl.NAryExpr(c.tok, new Bpl.FunctionCall(GetReadonlyField(f)), tyexprs);
-        var is_hf = MkIs(oDotF, f.Type);              // $Is(h[o, f], ..)
-        Bpl.Expr ax = bvsTypeAxiom.Count == 0 ? is_hf : BplForall(bvsTypeAxiom, BplTrigger(oDotF), is_hf);
-        sink.AddTopLevelDeclaration(new Bpl.Axiom(c.tok, BplImp(heightAntecedent, ax), string.Format("{0}.{1}: Type axiom", c, f)));
-
-        if (CommonHeapUse || (NonGhostsUseHeap && !f.IsGhost)) {
-          Bpl.Expr h;
-          var hVar = BplBoundVar("$h", predef.HeapType, out h);
-          bvsAllocationAxiom.Add(hVar);
-          var isGoodHeap = FunctionCall(c.tok, BuiltinFunction.IsGoodHeap, null, h);
-          var isalloc_hf = MkIsAlloc(oDotF, f.Type, h); // $IsAlloc(h[o, f], ..)
-          ax = BplForall(bvsAllocationAxiom, BplTrigger(isalloc_hf), BplImp(isGoodHeap, isalloc_hf));
-          sink.AddTopLevelDeclaration(new Bpl.Axiom(c.tok, BplImp(heightAntecedent, ax), string.Format("{0}.{1}: Allocation axiom", c, f)));
-        }
-
-      } else {
-        // This is the typical case (that is, f is not a static const field)
-
-        // h, o
-        Bpl.Expr h, o;
-        var hVar = BplBoundVar("$h", predef.HeapType, out h);
-        var oVar = BplBoundVar("$o", TrType(Resolver.GetThisType(c.tok, c)), out o);
-
-        // TClassA(G)
-        Bpl.Expr o_ty = ClassTyCon(c, tyexprs);
-
-        var isGoodHeap = FunctionCall(c.tok, BuiltinFunction.IsGoodHeap, null, h);
-        Bpl.Expr isalloc_o;
-        if (!(c is ClassDecl)) {
-          var udt = UserDefinedType.FromTopLevelDecl(c.tok, c);
-          isalloc_o = MkIsAlloc(o, udt, h);
-        } else if (RevealedInScope(c)) {
-          isalloc_o = IsAlloced(c.tok, h, o);
-        } else {
-          // c is only provided, not revealed, in the scope. Use the non-null type decl's internal synonym
-          var cl = (ClassDecl)c;
-          Contract.Assert(cl.NonNullTypeDecl != null);
-          var udt = UserDefinedType.FromTopLevelDecl(c.tok, cl.NonNullTypeDecl);
-          isalloc_o = MkIsAlloc(o, udt, h);
-        }
-
-        Bpl.Expr indexBounds = Bpl.Expr.True;
-        Bpl.Expr oDotF;
-        if (is_array) {
-          // generate h[o,Index(ii)]
-          bvsTypeAxiom.Add(hVar); bvsTypeAxiom.Add(oVar);
-          bvsAllocationAxiom.Add(hVar); bvsAllocationAxiom.Add(oVar);
-
-          var ac = (ArrayClassDecl)c;
-          var ixs = new List<Bpl.Expr>();
-          for (int i = 0; i < ac.Dims; i++) {
-            Bpl.Expr e; Bpl.Variable v = BplBoundVar("$i" + i, Bpl.Type.Int, out e);
-            ixs.Add(e);
-            bvsTypeAxiom.Add(v);
-            bvsAllocationAxiom.Add(v);
-          }
-
-          oDotF = ReadHeap(c.tok, h, o, GetArrayIndexFieldName(c.tok, ixs));
-
-          for (int i = 0; i < ac.Dims; i++) {
-            // 0 <= i && i < _System.array.Length(o)
-            var e1 = Bpl.Expr.Le(Bpl.Expr.Literal(0), ixs[i]);
-            var ff = GetReadonlyField((Field)(ac.Members[i]));
-            var e2 = Bpl.Expr.Lt(ixs[i], new Bpl.NAryExpr(c.tok, new Bpl.FunctionCall(ff), new List<Bpl.Expr> { o }));
-            indexBounds = BplAnd(indexBounds, BplAnd(e1, e2));
-          }
-        } else if (f.IsMutable) {
-          // generate h[o,f]
-          oDotF = ReadHeap(c.tok, h, o, new Bpl.IdentifierExpr(c.tok, GetField(f)));
-          bvsTypeAxiom.Add(hVar); bvsTypeAxiom.Add(oVar);
-          bvsAllocationAxiom.Add(hVar); bvsAllocationAxiom.Add(oVar);
-        } else {
-          // generate f(G,o)
-          var args = new List<Bpl.Expr> { o };
-          if (f is ConstantField) {
-            args = Concat(tyexprs, args);
-          }
-          oDotF = new Bpl.NAryExpr(c.tok, new Bpl.FunctionCall(GetReadonlyField(f)), args);
-          bvsTypeAxiom.Add(oVar);
-          bvsAllocationAxiom.Add(hVar); bvsAllocationAxiom.Add(oVar);
-        }
-
-        // antecedent: some subset of: $IsHeap(h) && o != null && $Is(o, TClassA(G)) && indexBounds
-        Bpl.Expr ante = Bpl.Expr.True;
-        if (is_array || f.IsMutable) {
-          ante = BplAnd(ante, isGoodHeap);
-          // Note: for the allocation axiom, isGoodHeap is added back in for !f.IsMutable below
-        }
-        if (!(f is ConstantField)) {
-          Bpl.Expr is_o = BplAnd(
-            ReceiverNotNull(o),
-            c is TraitDecl ? MkIs(o, o_ty) : DType(o, o_ty));  // $Is(o, ..)  or  dtype(o) == o_ty
-          ante = BplAnd(ante, is_o);
-        }
-        ante = BplAnd(ante, indexBounds);
-
-        // trigger
-        var t_es = new List<Bpl.Expr>();
-        t_es.Add(oDotF);
-        if (tyvars.Count > 0 && (is_array || !(f is ConstantField))) {
-          t_es.Add(o_ty);
-        }
-        var tr = new Bpl.Trigger(c.tok, true, t_es);
-
-        // Now for the conclusion of the axioms
-        Bpl.Expr is_hf, isalloc_hf = null;
-        if (is_array) {
-          is_hf = MkIs(oDotF, tyexprs[0], true);
-          if (CommonHeapUse || NonGhostsUseHeap) {
-            isalloc_hf = MkIsAlloc(oDotF, tyexprs[0], h, true);
-          }
-        } else {
-          is_hf = MkIs(oDotF, f.Type);              // $Is(h[o, f], ..)
-          if (CommonHeapUse || (NonGhostsUseHeap && !f.IsGhost)) {
-            isalloc_hf = MkIsAlloc(oDotF, f.Type, h); // $IsAlloc(h[o, f], ..)
-          }
-        }
-
-        Bpl.Expr ax = BplForall(bvsTypeAxiom, tr, BplImp(ante, is_hf));
-        sink.AddTopLevelDeclaration(new Bpl.Axiom(c.tok, BplImp(heightAntecedent, ax), string.Format("{0}.{1}: Type axiom", c, f)));
-
-        if (isalloc_hf != null) {
-          if (!is_array && !f.IsMutable) {
-            // isGoodHeap wasn't added above, so add it now
-            ante = BplAnd(isGoodHeap, ante);
-          }
-          ante = BplAnd(ante, isalloc_o);
-
-          // compute a different trigger
-          t_es = new List<Bpl.Expr>();
-          t_es.Add(oDotF);
-          if (!is_array && !f.IsMutable) {
-            // since "h" is not part of oDotF, we add a separate term that mentions "h"
-            t_es.Add(isalloc_o);
-          }
-          if (!(f is ConstantField) && tyvars.Count > 0) {
-            t_es.Add(o_ty);
-          }
-          tr = new Bpl.Trigger(c.tok, true, t_es);
-
-          ax = BplForall(bvsAllocationAxiom, tr, BplImp(ante, isalloc_hf));
-          sink.AddTopLevelDeclaration(new Bpl.Axiom(c.tok, BplImp(heightAntecedent, ax), string.Format("{0}.{1}: Allocation axiom", c, f)));
-        }
       }
     }
 
@@ -4318,222 +3125,6 @@ namespace Microsoft.Dafny {
       } else {
         return expr;
       }
-    }
-
-    void AddMethodImpl(Method m, Bpl.Procedure proc, bool wellformednessProc) {
-      Contract.Requires(m != null);
-      Contract.Requires(proc != null);
-      Contract.Requires(sink != null && predef != null);
-      Contract.Requires(wellformednessProc || m.Body != null);
-      Contract.Requires(currentModule == null && codeContext == null && _tmpIEs.Count == 0 && isAllocContext == null);
-      Contract.Ensures(currentModule == null && codeContext == null && _tmpIEs.Count == 0 && isAllocContext == null);
-
-      currentModule = m.EnclosingClass.EnclosingModuleDefinition;
-      codeContext = m;
-      isAllocContext = new IsAllocContext(m.IsGhost);
-
-      List<Variable> inParams = Bpl.Formal.StripWhereClauses(proc.InParams);
-      List<Variable> outParams = Bpl.Formal.StripWhereClauses(proc.OutParams);
-
-      BoogieStmtListBuilder builder = new BoogieStmtListBuilder(this);
-      builder.Add(new CommentCmd("AddMethodImpl: " + m + ", " + proc));
-      var etran = new ExpressionTranslator(this, predef, m.tok);
-      InitializeFuelConstant(m.tok, builder, etran);
-      var localVariables = new List<Variable>();
-      GenerateImplPrelude(m, wellformednessProc, inParams, outParams, builder, localVariables);
-
-      if (UseOptimizationInZ3) {
-        // We ask Z3 to minimize all parameters of type 'nat'.
-        foreach (var f in m.Ins) {
-          var udt = f.Type.NormalizeExpandKeepConstraints() as UserDefinedType;
-          if (udt != null && udt.Name == "nat") {
-            builder.Add(optimizeExpr(true, new IdentifierExpr(f.tok, f), f.Tok, etran));
-          }
-        }
-      }
-
-      Bpl.StmtList stmts;
-      if (!wellformednessProc) {
-        var inductionVars = ApplyInduction(m.Ins, m.Attributes);
-        if (inductionVars.Count != 0) {
-          // Let the parameters be this,x,y of the method M and suppose ApplyInduction returns y.
-          // Also, let Pre be the precondition and VF be the decreases clause.
-          // Then, insert into the method body what amounts to:
-          //     assume case-analysis-on-parameter[[ y' ]];
-          //     forall (y' | Pre(this, x, y') && VF(this, x, y') << VF(this, x, y)) {
-          //       this.M(x, y');
-          //     }
-          // Generate bound variables for the forall statement, and a substitution for the Pre and VF
-
-          // assume case-analysis-on-parameter[[ y' ]];
-          foreach (var inFormal in m.Ins) {
-            var dt = inFormal.Type.AsDatatype;
-            if (dt != null) {
-              var funcID = new Bpl.FunctionCall(new Bpl.IdentifierExpr(inFormal.tok, "$IsA#" + dt.FullSanitizedName, Bpl.Type.Bool));
-              var f = new Bpl.IdentifierExpr(inFormal.tok, inFormal.AssignUniqueName(m.IdGenerator), TrType(inFormal.Type));
-              builder.Add(TrAssumeCmd(inFormal.tok, new Bpl.NAryExpr(inFormal.tok, funcID, new List<Bpl.Expr> { f })));
-            }
-          }
-
-          var parBoundVars = new List<BoundVar>();
-          var parBounds = new List<ComprehensionExpr.BoundedPool>();
-          var substMap = new Dictionary<IVariable, Expression>();
-          Expression receiverSubst = null;
-          foreach (var iv in inductionVars) {
-            BoundVar bv;
-            if (iv == null) {
-              // this corresponds to "this"
-              Contract.Assert(!m.IsStatic);  // if "m" is static, "this" should never have gone into the _induction attribute
-              Contract.Assert(receiverSubst == null);  // we expect at most one
-              var receiverType = Resolver.GetThisType(m.tok, (TopLevelDeclWithMembers)m.EnclosingClass);
-              bv = new BoundVar(m.tok, CurrentIdGenerator.FreshId("$ih#this"), receiverType); // use this temporary variable counter, but for a Dafny name (the idea being that the number and the initial "_" in the name might avoid name conflicts)
-              var ie = new IdentifierExpr(m.tok, bv.Name);
-              ie.Var = bv;  // resolve here
-              ie.Type = bv.Type;  // resolve here
-              receiverSubst = ie;
-            } else {
-              IdentifierExpr ie;
-              CloneVariableAsBoundVar(iv.tok, iv, "$ih#" + iv.Name, out bv, out ie);
-              substMap.Add(iv, ie);
-            }
-            parBoundVars.Add(bv);
-            parBounds.Add(new ComprehensionExpr.SpecialAllocIndependenceAllocatedBoundedPool());  // record that we don't want alloc antecedents for these variables
-          }
-
-          // Generate a CallStmt for the recursive call
-          Expression recursiveCallReceiver;
-          List<Expression> recursiveCallArgs;
-          RecursiveCallParameters(m.tok, m, m.TypeArgs, m.Ins, substMap, out recursiveCallReceiver, out recursiveCallArgs);
-          var methodSel = new MemberSelectExpr(m.tok, recursiveCallReceiver, m.Name);
-          methodSel.Member = m;  // resolve here
-          methodSel.TypeApplication_AtEnclosingClass = m.EnclosingClass.TypeArgs.ConvertAll(tp => (Type)new UserDefinedType(tp.tok, tp));
-          methodSel.TypeApplication_JustMember = m.TypeArgs.ConvertAll(tp => (Type)new UserDefinedType(tp.tok, tp));
-          methodSel.Type = new InferredTypeProxy();
-          var recursiveCall = new CallStmt(m.tok, m.tok, new List<Expression>(), methodSel, recursiveCallArgs);
-          recursiveCall.IsGhost = m.IsGhost;  // resolve here
-
-          Expression parRange = new LiteralExpr(m.tok, true);
-          parRange.Type = Type.Bool;  // resolve here
-          foreach (var pre in m.Req) {
-            parRange = Expression.CreateAnd(parRange, Substitute(pre.E, receiverSubst, substMap));
-          }
-          // construct an expression (generator) for:  VF' << VF
-          ExpressionConverter decrCheck = delegate (Dictionary<IVariable, Expression> decrSubstMap, ExpressionTranslator exprTran) {
-            var decrToks = new List<IToken>();
-            var decrTypes = new List<Type>();
-            var decrCallee = new List<Expr>();
-            var decrCaller = new List<Expr>();
-            foreach (var ee in m.Decreases.Expressions) {
-              decrToks.Add(ee.tok);
-              decrTypes.Add(ee.Type.NormalizeExpand());
-              decrCaller.Add(exprTran.TrExpr(ee));
-              Expression es = Substitute(ee, receiverSubst, substMap);
-              es = Substitute(es, null, decrSubstMap);
-              decrCallee.Add(exprTran.TrExpr(es));
-            }
-            return DecreasesCheck(decrToks, decrTypes, decrTypes, decrCallee, decrCaller, null, null, false, true);
-          };
-
-#if VERIFY_CORRECTNESS_OF_TRANSLATION_FORALL_STATEMENT_RANGE
-          var definedness = new BoogieStmtListBuilder(this);
-          var exporter = new BoogieStmtListBuilder(this);
-          TrForallStmtCall(m.tok, parBoundVars, parRange, decrCheck, null, recursiveCall, definedness, exporter, localVariables, etran);
-          // All done, so put the two pieces together
-          builder.Add(new Bpl.IfCmd(m.tok, null, definedness.Collect(m.tok), null, exporter.Collect(m.tok)));
-#else
-          TrForallStmtCall(m.tok, parBoundVars, parBounds, parRange, decrCheck, null, recursiveCall, null, builder, localVariables, etran);
-#endif
-        }
-        // translate the body of the method
-        Contract.Assert(m.Body != null);  // follows from method precondition and the if guard
-
-        // $_reverifyPost := false;
-        builder.Add(Bpl.Cmd.SimpleAssign(m.tok, new Bpl.IdentifierExpr(m.tok, "$_reverifyPost", Bpl.Type.Bool), Bpl.Expr.False));
-        // register output parameters with definite-assignment trackers
-        Contract.Assert(definiteAssignmentTrackers.Count == 0);
-        m.Outs.Iter(p => AddExistingDefiniteAssignmentTracker(p, m.IsGhost));
-        // translate the body
-        TrStmt(m.Body, builder, localVariables, etran);
-        m.Outs.Iter(p => CheckDefiniteAssignmentReturn(m.BodyEndTok, p, builder));
-        stmts = builder.Collect(m.Body.Tok);
-        // tear down definite-assignment trackers
-        m.Outs.Iter(RemoveDefiniteAssignmentTracker);
-        Contract.Assert(definiteAssignmentTrackers.Count == 0);
-      } else {
-        // check well-formedness of any default-value expressions (before assuming preconditions)
-        foreach (var formal in m.Ins.Where(formal => formal.DefaultValue != null)) {
-          var e = formal.DefaultValue;
-          CheckWellformed(e, new WFOptions(null, false, false, true), localVariables, builder, etran);
-          builder.Add(new Bpl.AssumeCmd(e.tok, CanCallAssumption(e, etran)));
-          CheckSubrange(e.tok, etran.TrExpr(e), e.Type, formal.Type, builder);
-
-          if (formal.IsOld) {
-            Bpl.Expr wh = GetWhereClause(e.tok, etran.TrExpr(e), e.Type, etran.Old, ISALLOC, true);
-            if (wh != null) {
-              builder.Add(Assert(e.tok, wh, "default value must be allocated in the two-state lemma's previous state"));
-            }
-          }
-        }
-        // check well-formedness of the preconditions, and then assume each one of them
-        foreach (AttributedExpression p in m.Req) {
-          CheckWellformedAndAssume(p.E, new WFOptions(), localVariables, builder, etran);
-        }
-        // check well-formedness of the modifies clauses
-        CheckFrameWellFormed(new WFOptions(), m.Mod.Expressions, localVariables, builder, etran);
-        // check well-formedness of the decreases clauses
-        foreach (Expression p in m.Decreases.Expressions) {
-          CheckWellformed(p, new WFOptions(), localVariables, builder, etran);
-        }
-
-        if (!(m is TwoStateLemma)) {
-          // play havoc with the heap according to the modifies clause
-          builder.Add(new Bpl.HavocCmd(m.tok, new List<Bpl.IdentifierExpr> { (Bpl.IdentifierExpr/*TODO: this cast is rather dubious*/)etran.HeapExpr }));
-          // assume the usual two-state boilerplate information
-          foreach (BoilerplateTriple tri in GetTwoStateBoilerplate(m.tok, m.Mod.Expressions, m.IsGhost, m.AllowsAllocation, etran.Old, etran, etran.Old)) {
-            if (tri.IsFree) {
-              builder.Add(TrAssumeCmd(m.tok, tri.Expr));
-            }
-          }
-        }
-
-        // also play havoc with the out parameters
-        if (outParams.Count != 0) {  // don't create an empty havoc statement
-          List<Bpl.IdentifierExpr> outH = new List<Bpl.IdentifierExpr>();
-          foreach (Bpl.Variable b in outParams) {
-            Contract.Assert(b != null);
-            outH.Add(new Bpl.IdentifierExpr(b.tok, b));
-          }
-          builder.Add(new Bpl.HavocCmd(m.tok, outH));
-        }
-        // mark the end of the modifles/out-parameter havocking with a CaptureState; make its location be the first ensures clause, if any (and just
-        // omit the CaptureState if there's no ensures clause)
-        if (m.Ens.Count != 0) {
-          builder.AddCaptureState(m.Ens[0].E.tok, false, "post-state");
-        }
-
-        // check wellformedness of postconditions
-        foreach (AttributedExpression p in m.Ens) {
-          CheckWellformedAndAssume(p.E, new WFOptions(), localVariables, builder, etran);
-        }
-
-        stmts = builder.Collect(m.tok);
-      }
-
-      if (EmitImplementation(m.Attributes)) {
-        // emit impl only when there are proof obligations.
-        QKeyValue kv = etran.TrAttributes(m.Attributes, null);
-        Bpl.Implementation impl = new Bpl.Implementation(m.tok, proc.Name,
-          new List<Bpl.TypeVariable>(), inParams, outParams,
-          localVariables, stmts, kv);
-        sink.AddTopLevelDeclaration(impl);
-
-        if (InsertChecksums) {
-          InsertChecksum(m, impl);
-        }
-      }
-
-      isAllocContext = null;
-      Reset();
     }
 
     #region Definite-assignment tracking
@@ -4752,187 +3343,6 @@ namespace Microsoft.Dafny {
       return assumeCmd;
     }
 
-    private void AddFunctionOverrideCheckImpl(Function f) {
-      Contract.Requires(f != null);
-      Contract.Requires(f.EnclosingClass is TopLevelDeclWithMembers);
-      Contract.Requires(sink != null && predef != null);
-      Contract.Requires(f.OverriddenFunction != null);
-      Contract.Requires(f.Formals.Count == f.OverriddenFunction.Formals.Count);
-      Contract.Requires(currentModule == null && codeContext == null && _tmpIEs.Count == 0 && isAllocContext != null);
-      Contract.Ensures(currentModule == null && codeContext == null && _tmpIEs.Count == 0 && isAllocContext != null);
-
-      #region first procedure, no impl yet
-      //Function nf = new Function(f.tok, "OverrideCheck_" + f.Name, f.IsStatic, f.IsGhost, f.TypeArgs, f.OpenParen, f.Formals, f.ResultType, f.Req, f.Reads, f.Ens, f.Decreases, f.Body, f.Attributes, f.SignatureEllipsis);
-      //AddFunction(f);
-      currentModule = f.EnclosingClass.EnclosingModuleDefinition;
-      codeContext = f;
-
-      Bpl.Expr prevHeap = null;
-      Bpl.Expr currHeap = null;
-      var ordinaryEtran = new ExpressionTranslator(this, predef, f.tok);
-      ExpressionTranslator etran;
-      var inParams_Heap = new List<Bpl.Variable>();
-      if (f is TwoStateFunction) {
-        var prevHeapVar = new Bpl.Formal(f.tok, new Bpl.TypedIdent(f.tok, "previous$Heap", predef.HeapType), true);
-        inParams_Heap.Add(prevHeapVar);
-        prevHeap = new Bpl.IdentifierExpr(f.tok, prevHeapVar);
-        if (f.ReadsHeap) {
-          var currHeapVar = new Bpl.Formal(f.tok, new Bpl.TypedIdent(f.tok, "current$Heap", predef.HeapType), true);
-          inParams_Heap.Add(currHeapVar);
-          currHeap = new Bpl.IdentifierExpr(f.tok, currHeapVar);
-        }
-        etran = new ExpressionTranslator(this, predef, currHeap, prevHeap);
-      } else {
-        etran = ordinaryEtran;
-      }
-
-      // parameters of the procedure
-      var typeInParams = MkTyParamFormals(GetTypeParams(f));
-      var inParams = new List<Variable>();
-      var outParams = new List<Bpl.Variable>();
-      if (!f.IsStatic) {
-        var th = new Bpl.IdentifierExpr(f.tok, "this", TrReceiverType(f));
-        Bpl.Expr wh = Bpl.Expr.And(
-          ReceiverNotNull(th),
-          etran.GoodRef(f.tok, th, Resolver.GetReceiverType(f.tok, f)));
-        Bpl.Formal thVar = new Bpl.Formal(f.tok, new Bpl.TypedIdent(f.tok, "this", TrReceiverType(f), wh), true);
-        inParams.Add(thVar);
-      }
-      foreach (Formal p in f.Formals) {
-        Bpl.Type varType = TrType(p.Type);
-        Bpl.Expr wh = GetWhereClause(p.tok, new Bpl.IdentifierExpr(p.tok, p.AssignUniqueName(f.IdGenerator), varType), p.Type, etran, NOALLOC);
-        inParams.Add(new Bpl.Formal(p.tok, new Bpl.TypedIdent(p.tok, p.AssignUniqueName(f.IdGenerator), varType, wh), true));
-      }
-
-      Formal pOut = null;
-      if (f.Result != null || f.OverriddenFunction.Result != null) {
-        if (f.Result != null) {
-          pOut = f.Result;
-          Contract.Assert(!pOut.IsOld);
-        } else {
-          var pp = f.OverriddenFunction.Result;
-          Contract.Assert(!pp.IsOld);
-          pOut = new Formal(pp.tok, pp.Name, f.ResultType, false, pp.IsGhost, null);
-        }
-        var varType = TrType(pOut.Type);
-        var wh = GetWhereClause(pOut.tok, new Bpl.IdentifierExpr(pOut.tok, pOut.AssignUniqueName(f.IdGenerator), varType), pOut.Type, etran, NOALLOC);
-        outParams.Add(new Bpl.Formal(pOut.tok, new Bpl.TypedIdent(pOut.tok, pOut.AssignUniqueName(f.IdGenerator), varType, wh), true));
-      }
-      // the procedure itself
-      var req = new List<Bpl.Requires>();
-      // free requires mh == ModuleContextHeight && fh == FunctionContextHeight;
-      req.Add(Requires(f.tok, true, etran.HeightContext(f.OverriddenFunction), null, null));
-      if (f is TwoStateFunction) {
-        // free requires prevHeap == Heap && HeapSucc(prevHeap, currHeap) && IsHeap(currHeap)
-        var a0 = Bpl.Expr.Eq(prevHeap, ordinaryEtran.HeapExpr);
-        var a1 = HeapSucc(prevHeap, currHeap);
-        var a2 = FunctionCall(f.tok, BuiltinFunction.IsGoodHeap, null, currHeap);
-        req.Add(Requires(f.tok, true, BplAnd(a0, BplAnd(a1, a2)), null, null));
-      }
-      // modifies $Heap, $Tick
-      var mod = new List<Bpl.IdentifierExpr> {
-          (Bpl.IdentifierExpr/*TODO: this cast is rather dubious*/)ordinaryEtran.HeapExpr,
-          etran.Tick()
-        };
-      var ens = new List<Bpl.Ensures>();
-
-      var proc = new Bpl.Procedure(f.tok, "OverrideCheck$$" + f.FullSanitizedName, new List<Bpl.TypeVariable>(),
-        Concat(Concat(typeInParams, inParams_Heap), inParams), outParams,
-        req, mod, ens, etran.TrAttributes(f.Attributes, null));
-      sink.AddTopLevelDeclaration(proc);
-      var implInParams = Bpl.Formal.StripWhereClauses(inParams);
-      var implOutParams = Bpl.Formal.StripWhereClauses(outParams);
-
-      #endregion
-
-      //List<Variable> outParams = Bpl.Formal.StripWhereClauses(proc.OutParams);
-
-      BoogieStmtListBuilder builder = new BoogieStmtListBuilder(this);
-      List<Variable> localVariables = new List<Variable>();
-
-      // assume traitTypeParameter == G(overrideTypeParameters);
-      AddOverrideCheckTypeArgumentInstantiations(f, builder, localVariables);
-
-      if (f is TwoStateFunction) {
-        // $Heap := current$Heap;
-        var heap = (Bpl.IdentifierExpr /*TODO: this cast is somewhat dubious*/)ordinaryEtran.HeapExpr;
-        builder.Add(Bpl.Cmd.SimpleAssign(f.tok, heap, etran.HeapExpr));
-        etran = ordinaryEtran;  // we no longer need the special heap names
-      }
-
-      var substMap = new Dictionary<IVariable, Expression>();
-      for (int i = 0; i < f.Formals.Count; i++) {
-        //get corresponsing formal in the class
-        var ie = new IdentifierExpr(f.Formals[i].tok, f.Formals[i].AssignUniqueName(f.IdGenerator));
-        ie.Var = f.Formals[i]; ie.Type = ie.Var.Type;
-        substMap.Add(f.OverriddenFunction.Formals[i], ie);
-      }
-
-      if (f.OverriddenFunction.Result != null) {
-        Contract.Assert(pOut != null);
-        //get corresponsing formal in the class
-        var ie = new IdentifierExpr(pOut.tok, pOut.AssignUniqueName(f.IdGenerator));
-        ie.Var = pOut; ie.Type = ie.Var.Type;
-        substMap.Add(f.OverriddenFunction.Result, ie);
-      }
-
-      //adding assume Pre’; assert P; // this checks that Pre’ implies P
-      AddFunctionOverrideReqsChk(f, builder, etran, substMap);
-
-      //adding assert R <= Rank’;
-      AddOverrideTerminationChk(f, f.OverriddenFunction, builder, etran, substMap);
-
-      //adding assert W <= Frame’
-      AddFunctionOverrideSubsetChk(f, builder, etran, localVariables, substMap);
-
-      //adding assume Q; assert Post’;
-      //adding assume J.F(ins) == C.F(ins);
-      AddFunctionOverrideEnsChk(f, builder, etran, substMap, implInParams, implOutParams.Count == 0 ? null : implOutParams[0]);
-
-      var stmts = builder.Collect(f.tok);
-
-      if (EmitImplementation(f.Attributes)) {
-        // emit the impl only when there are proof obligations.
-        QKeyValue kv = etran.TrAttributes(f.Attributes, null);
-
-        var impl = new Bpl.Implementation(f.tok, proc.Name, new List<Bpl.TypeVariable>(),
-          Concat(Concat(typeInParams, inParams_Heap), implInParams), implOutParams, localVariables, stmts, kv);
-        sink.AddTopLevelDeclaration(impl);
-      }
-
-      if (InsertChecksums) {
-        InsertChecksum(f, proc, true);
-      }
-
-      Reset();
-    }
-
-    private void AddOverrideCheckTypeArgumentInstantiations(MemberDecl member, BoogieStmtListBuilder builder, List<Variable> localVariables) {
-      Contract.Requires(member is Function || member is Method);
-      Contract.Requires(member.EnclosingClass is TopLevelDeclWithMembers);
-      Contract.Requires(builder != null);
-      Contract.Requires(localVariables != null);
-
-      MemberDecl overriddenMember;
-      List<TypeParameter> overriddenTypeParameters;
-      if (member is Function) {
-        var o = ((Function)member).OverriddenFunction;
-        overriddenMember = o;
-        overriddenTypeParameters = o.TypeArgs;
-      } else {
-        var o = ((Method)member).OverriddenMethod;
-        overriddenMember = o;
-        overriddenTypeParameters = o.TypeArgs;
-      }
-      var typeMap = GetTypeArgumentSubstitutionMap(overriddenMember, member);
-      foreach (var tp in Concat(overriddenMember.EnclosingClass.TypeArgs, overriddenTypeParameters)) {
-        var local = BplLocalVar(nameTypeParam(tp), predef.Ty, out var lhs);
-        localVariables.Add(local);
-        var rhs = TypeToTy(typeMap[tp]);
-        builder.Add(new Bpl.AssumeCmd(tp.tok, Bpl.Expr.Eq(lhs, rhs)));
-      }
-    }
-
     private void AddFunctionOverrideEnsChk(Function f, BoogieStmtListBuilder builder, ExpressionTranslator etran, Dictionary<IVariable, Expression> substMap, List<Bpl.Variable> implInParams, Bpl.Variable/*?*/ resultVariable) {
       Contract.Requires(f.Formals.Count <= implInParams.Count);
 
@@ -5040,47 +3450,6 @@ namespace Microsoft.Dafny {
       return tyargs;
     }
 
-    /// <summary>
-    /// Return a type-parameter substitution map for function "f", as instantiated by the context of "overridingFunction".
-    ///
-    /// In more symbols, suppose "f" is declared as follows:
-    ///     class/trait Tr[A,B] {
-    ///       function f[C,D](...): ...
-    ///     }
-    /// and "overridingFunction" is declared as follows:
-    ///     class/trait Cl[G] extends Tr[X(G),Y(G)] {
-    ///       function f[R,S](...): ...
-    ///     }
-    /// Then, return the following map:
-    ///     A -> X(G)
-    ///     B -> Y(G)
-    ///     C -> R
-    ///     D -> S
-    ///
-    /// See also GetTypeArguments.
-    /// </summary>
-    private static Dictionary<TypeParameter, Type> GetTypeArgumentSubstitutionMap(MemberDecl member, MemberDecl overridingMember) {
-      Contract.Requires(member is Function || member is Method);
-      Contract.Requires(overridingMember is Function || overridingMember is Method);
-      Contract.Requires(overridingMember.EnclosingClass is TopLevelDeclWithMembers);
-      Contract.Requires(((ICallable)member).TypeArgs.Count == ((ICallable)overridingMember).TypeArgs.Count);
-
-      var typeMap = new Dictionary<TypeParameter, Type>();
-
-      var cl = (TopLevelDeclWithMembers)overridingMember.EnclosingClass;
-      var classTypeMap = cl.ParentFormalTypeParametersToActuals;
-      member.EnclosingClass.TypeArgs.ForEach(tp => typeMap.Add(tp, classTypeMap[tp]));
-
-      var origTypeArgs = ((ICallable)member).TypeArgs;
-      var overridingTypeArgs = ((ICallable)overridingMember).TypeArgs;
-      for (var i = 0; i < origTypeArgs.Count; i++) {
-        var otp = overridingTypeArgs[i];
-        typeMap.Add(origTypeArgs[i], new UserDefinedType(otp.tok, otp));
-      }
-
-      return typeMap;
-    }
-
     private void AddFunctionOverrideSubsetChk(Function func, BoogieStmtListBuilder builder, ExpressionTranslator etran, List<Variable> localVariables, Dictionary<IVariable, Expression> substMap) {
       //getting framePrime
       List<FrameExpression> traitFrameExps = new List<FrameExpression>();
@@ -5142,87 +3511,6 @@ namespace Microsoft.Dafny {
       }
     }
 
-    private void AddMethodOverrideCheckImpl(Method m, Bpl.Procedure proc) {
-      Contract.Requires(m != null);
-      Contract.Requires(proc != null);
-      Contract.Requires(sink != null && predef != null);
-      Contract.Requires(m.OverriddenMethod != null);
-      Contract.Requires(m.Ins.Count == m.OverriddenMethod.Ins.Count);
-      Contract.Requires(m.Outs.Count == m.OverriddenMethod.Outs.Count);
-      //Contract.Requires(wellformednessProc || m.Body != null);
-      Contract.Requires(currentModule == null && codeContext == null && _tmpIEs.Count == 0 && isAllocContext == null);
-      Contract.Ensures(currentModule == null && codeContext == null && _tmpIEs.Count == 0 && isAllocContext == null);
-
-      currentModule = m.EnclosingClass.EnclosingModuleDefinition;
-      codeContext = m;
-      isAllocContext = new IsAllocContext(m.IsGhost);
-
-      List<Variable> inParams = Bpl.Formal.StripWhereClauses(proc.InParams);
-      List<Variable> outParams = Bpl.Formal.StripWhereClauses(proc.OutParams);
-
-      var builder = new BoogieStmtListBuilder(this);
-      var etran = new ExpressionTranslator(this, predef, m.tok);
-      var localVariables = new List<Variable>();
-
-      // assume traitTypeParameter == G(overrideTypeParameters);
-      AddOverrideCheckTypeArgumentInstantiations(m, builder, localVariables);
-
-      if (m is TwoStateLemma) {
-        // $Heap := current$Heap;
-        var heap = (Bpl.IdentifierExpr /*TODO: this cast is somewhat dubious*/)new ExpressionTranslator(this, predef, m.tok).HeapExpr;
-        builder.Add(Bpl.Cmd.SimpleAssign(m.tok, heap, new Bpl.IdentifierExpr(m.tok, "current$Heap", predef.HeapType)));
-      }
-
-
-      var substMap = new Dictionary<IVariable, Expression>();
-      for (int i = 0; i < m.Ins.Count; i++) {
-        //get corresponsing formal in the class
-        var ie = new IdentifierExpr(m.Ins[i].tok, m.Ins[i].AssignUniqueName(m.IdGenerator));
-        ie.Var = m.Ins[i]; ie.Type = ie.Var.Type;
-        substMap.Add(m.OverriddenMethod.Ins[i], ie);
-      }
-      for (int i = 0; i < m.Outs.Count; i++) {
-        //get corresponsing formal in the class
-        var ie = new IdentifierExpr(m.Outs[i].tok, m.Outs[i].AssignUniqueName(m.IdGenerator));
-        ie.Var = m.Outs[i]; ie.Type = ie.Var.Type;
-        substMap.Add(m.OverriddenMethod.Outs[i], ie);
-      }
-
-      Bpl.StmtList stmts;
-      //adding assume Pre’; assert P; // this checks that Pre’ implies P
-      AddMethodOverrideReqsChk(m, builder, etran, substMap);
-
-      //adding assert R <= Rank’;
-      AddOverrideTerminationChk(m, m.OverriddenMethod, builder, etran, substMap);
-
-      //adding assert W <= Frame’
-      AddMethodOverrideSubsetChk(m, builder, etran, localVariables, substMap);
-
-      if (!(m is TwoStateLemma)) {
-        //change the heap at locations W
-        HavocMethodFrameLocations(m, builder, etran, localVariables);
-      }
-
-      //adding assume Q; assert Post’;
-      AddMethodOverrideEnsChk(m, builder, etran, substMap);
-
-      stmts = builder.Collect(m.tok);
-
-      if (EmitImplementation(m.Attributes)) {
-        // emit the impl only when there are proof obligations.
-        QKeyValue kv = etran.TrAttributes(m.Attributes, null);
-        Bpl.Implementation impl = new Bpl.Implementation(m.tok, proc.Name, new List<Bpl.TypeVariable>(), inParams, outParams, localVariables, stmts, kv);
-        sink.AddTopLevelDeclaration(impl);
-
-        if (InsertChecksums) {
-          InsertChecksum(m, impl);
-        }
-      }
-
-      isAllocContext = null;
-      Reset();
-    }
-
     private void HavocMethodFrameLocations(Method m, BoogieStmtListBuilder builder, ExpressionTranslator etran, List<Variable> localVariables) {
       Contract.Requires(m != null);
       Contract.Requires(m.EnclosingClass != null && m.EnclosingClass is ClassDecl);
@@ -5235,181 +3523,6 @@ namespace Microsoft.Dafny {
           builder.Add(TrAssumeCmd(m.tok, tri.Expr));
         }
       }
-    }
-
-    private void AddMethodOverrideEnsChk(Method m, BoogieStmtListBuilder builder, ExpressionTranslator etran, Dictionary<IVariable, Expression> substMap) {
-      Contract.Requires(m != null);
-      Contract.Requires(builder != null);
-      Contract.Requires(etran != null);
-      Contract.Requires(substMap != null);
-      //generating class post-conditions
-      foreach (var en in m.Ens) {
-        builder.Add(TrAssumeCmd(m.tok, etran.TrExpr(en.E)));
-      }
-      //generating trait post-conditions with class variables
-      foreach (var en in m.OverriddenMethod.Ens) {
-        Expression postcond = Substitute(en.E, null, substMap);
-        bool splitHappened;  // we actually don't care
-        foreach (var s in TrSplitExpr(postcond, etran, false, out splitHappened)) {
-          if (s.IsChecked) {
-            builder.Add(Assert(m.tok, s.E, "the method must provide an equal or more detailed postcondition than in its parent trait"));
-          }
-        }
-      }
-    }
-
-    private void AddMethodOverrideReqsChk(Method m, BoogieStmtListBuilder builder, ExpressionTranslator etran, Dictionary<IVariable, Expression> substMap) {
-      Contract.Requires(m != null);
-      Contract.Requires(builder != null);
-      Contract.Requires(etran != null);
-      Contract.Requires(substMap != null);
-      //generating trait pre-conditions with class variables
-      foreach (var req in m.OverriddenMethod.Req) {
-        Expression precond = Substitute(req.E, null, substMap);
-        builder.Add(TrAssumeCmd(m.tok, etran.TrExpr(precond)));
-      }
-      //generating class pre-conditions
-      foreach (var req in m.Req) {
-        bool splitHappened;  // we actually don't care
-        foreach (var s in TrSplitExpr(req.E, etran, false, out splitHappened)) {
-          if (s.IsChecked) {
-            builder.Add(Assert(m.tok, s.E, "the method must provide an equal or more permissive precondition than in its parent trait"));
-          }
-        }
-      }
-    }
-
-    private void AddOverrideTerminationChk(ICallable original, ICallable overryd, BoogieStmtListBuilder builder, ExpressionTranslator etran, Dictionary<IVariable, Expression> substMap) {
-      Contract.Requires(original != null);
-      Contract.Requires(overryd != null);
-      Contract.Requires(builder != null);
-      Contract.Requires(etran != null);
-      Contract.Requires(substMap != null);
-      // Note, it is as if the trait's method is calling the class's method.
-      var contextDecreases = overryd.Decreases.Expressions;
-      var calleeDecreases = original.Decreases.Expressions;
-      // We want to check:  calleeDecreases <= contextDecreases (note, we can allow equality, since there is a bounded, namely 1, number of dynamic dispatches)
-      if (Contract.Exists(contextDecreases, e => e is WildcardExpr)) {
-        // no check needed
-        return;
-      }
-
-      int N = Math.Min(contextDecreases.Count, calleeDecreases.Count);
-      var toks = new List<IToken>();
-      var types0 = new List<Type>();
-      var types1 = new List<Type>();
-      var callee = new List<Expr>();
-      var caller = new List<Expr>();
-
-      for (int i = 0; i < N; i++) {
-        Expression e0 = calleeDecreases[i];
-        Expression e1 = Substitute(contextDecreases[i], null, substMap);
-        if (!CompatibleDecreasesTypes(e0.Type, e1.Type)) {
-          N = i;
-          break;
-        }
-        toks.Add(new NestedToken(original.Tok, e1.tok));
-        types0.Add(e0.Type.NormalizeExpand());
-        types1.Add(e1.Type.NormalizeExpand());
-        callee.Add(etran.TrExpr(e0));
-        caller.Add(etran.TrExpr(e1));
-      }
-
-      var decrCountT = contextDecreases.Count;
-      var decrCountC = calleeDecreases.Count;
-      // Generally, we want to produce a check "decrClass <= decrTrait", allowing (the common case where) they are equal.
-      // * If N < decrCountC && N < decrCountT, then "decrClass <= decrTrait" if the comparison ever gets beyond the
-      //   parts that survived truncation.  Thus, we compare with "allowNoChange" set to "false".
-      // Otherwise:
-      // * If decrCountC == decrCountT, then the truncation we did above had no effect and we pass in "allowNoChange" as "true".
-      // * If decrCountC > decrCountT, then we will have truncated decrClass above.  Let x,y and x' denote decrClass and
-      //   decrTrait, respectively, where x and x' have the same length.  Considering how Dafny in effect pads the end of
-      //   decreases tuples with a \top, we were supposed to evaluate (x,(y,\top)) <= (x',\top), which by lexicographic pairs
-      //   we can expand to:
-      //       x <= x' && (x == x' ==> (y,\top) <= \top)
-      //   which is equivalent to just x <= x'.  Thus, we called DecreasesCheck to compare x and x' and we pass in "allowNoChange"
-      //   as "true".
-      // * If decrCountC < decrCountT, then we will have truncated decrTrait above.  Let x and x',y' denote decrClass and
-      //   decrTrait, respectively, where x and x' have the same length.  We then want to check (x,\top) <= (x',(y',\top)), which
-      //   expands to:
-      //       x <= x' && (x == x' ==> \top <= (y',\top))
-      //    =      { \top is strictly larger than a pair }
-      //       x <= x' && (x == x' ==> false)
-      //    =
-      //       x < x'
-      //   So we perform our desired check by calling DecreasesCheck to strictly compare x and x', so we pass in "allowNoChange"
-      //   as "false".
-      bool allowNoChange = N == decrCountT && decrCountT <= decrCountC;
-      var decrChk = DecreasesCheck(toks, types0, types1, callee, caller, null, null, allowNoChange, false);
-      builder.Add(Assert(original.Tok, decrChk, string.Format("{0}'s decreases clause must be below or equal to that in the trait", original.WhatKind)));
-    }
-
-    private void AddMethodOverrideSubsetChk(Method m, BoogieStmtListBuilder builder, ExpressionTranslator etran, List<Variable> localVariables, Dictionary<IVariable, Expression> substMap) {
-      //getting framePrime
-      List<FrameExpression> traitFrameExps = new List<FrameExpression>();
-      List<FrameExpression> classFrameExps = m.Mod != null ? m.Mod.Expressions : new List<FrameExpression>();
-      if (m.OverriddenMethod.Mod != null) {
-        foreach (var e in m.OverriddenMethod.Mod.Expressions) {
-          var newE = Substitute(e.E, null, substMap);
-          FrameExpression fe = new FrameExpression(e.tok, newE, e.FieldName);
-          traitFrameExps.Add(fe);
-        }
-      }
-
-      QKeyValue kv = etran.TrAttributes(m.Attributes, null);
-
-      IToken tok = m.tok;
-      // Declare a local variable $_Frame: <alpha>[ref, Field alpha]bool
-      Bpl.IdentifierExpr traitFrame = etran.TheFrame(m.OverriddenMethod.tok);  // this is a throw-away expression, used only to extract the type and name of the $_Frame variable
-      traitFrame.Name = m.EnclosingClass.Name + "_" + traitFrame.Name;
-      Contract.Assert(traitFrame.Type != null);  // follows from the postcondition of TheFrame
-      Bpl.LocalVariable frame = new Bpl.LocalVariable(tok, new Bpl.TypedIdent(tok, null ?? traitFrame.Name, traitFrame.Type));
-      localVariables.Add(frame);
-      // $_Frame := (lambda<alpha> $o: ref, $f: Field alpha :: $o != null && $Heap[$o,alloc] ==> ($o,$f) in Modifies/Reads-Clause);
-      Bpl.TypeVariable alpha = new Bpl.TypeVariable(tok, "alpha");
-      Bpl.BoundVariable oVar = new Bpl.BoundVariable(tok, new Bpl.TypedIdent(tok, "$o", predef.RefType));
-      Bpl.IdentifierExpr o = new Bpl.IdentifierExpr(tok, oVar);
-      Bpl.BoundVariable fVar = new Bpl.BoundVariable(tok, new Bpl.TypedIdent(tok, "$f", predef.FieldName(tok, alpha)));
-      Bpl.IdentifierExpr f = new Bpl.IdentifierExpr(tok, fVar);
-      Bpl.Expr ante = Bpl.Expr.And(Bpl.Expr.Neq(o, predef.Null), etran.IsAlloced(tok, o));
-      Bpl.Expr consequent = InRWClause(tok, o, f, traitFrameExps, etran, null, null);
-      Bpl.Expr lambda = new Bpl.LambdaExpr(tok, new List<TypeVariable> { alpha }, new List<Variable> { oVar, fVar }, null,
-                                           Bpl.Expr.Imp(ante, consequent));
-
-      //to initialize $_Frame variable to Frame'
-      builder.Add(Bpl.Cmd.SimpleAssign(tok, new Bpl.IdentifierExpr(tok, frame), lambda));
-
-      // emit: assert (forall<alpha> o: ref, f: Field alpha :: o != null && $Heap[o,alloc] && (o,f) in subFrame ==> $_Frame[o,f]);
-      Bpl.Expr oInCallee = InRWClause(tok, o, f, classFrameExps, etran, null, null);
-      Bpl.Expr consequent2 = InRWClause(tok, o, f, traitFrameExps, etran, null, null);
-      Bpl.Expr q = new Bpl.ForallExpr(tok, new List<TypeVariable> { alpha }, new List<Variable> { oVar, fVar },
-                                      Bpl.Expr.Imp(Bpl.Expr.And(ante, oInCallee), consequent2));
-      builder.Add(Assert(tok, q, "expression may modify an object not in the parent trait context's modifies clause", kv));
-    }
-
-    private void InsertChecksum(Method m, Bpl.Declaration decl, bool specificationOnly = false) {
-      Contract.Requires(VisibleInScope(m));
-      byte[] data;
-      using (var writer = new System.IO.StringWriter()) {
-        var printer = new Printer(writer);
-        printer.PrintAttributes(m.Attributes);
-        printer.PrintFormals(m.Ins, m);
-        if (m.Outs.Any()) {
-          writer.Write("returns ");
-          printer.PrintFormals(m.Outs, m);
-        }
-        printer.PrintSpec("", m.Req, 0);
-        printer.PrintFrameSpecLine("", m.Mod.Expressions, 0, null);
-        printer.PrintSpec("", m.Ens, 0);
-        printer.PrintDecreasesSpec(m.Decreases, 0);
-        writer.WriteLine();
-        if (!specificationOnly && m.Body != null && RevealedInScope(m)) {
-          printer.PrintStatement(m.Body, 0);
-        }
-        data = Encoding.UTF8.GetBytes(writer.ToString());
-      }
-
-      InsertChecksum(decl, data);
     }
 
     private void InsertChecksum(DatatypeDecl d, Bpl.Declaration decl) {
@@ -8654,12 +6767,13 @@ namespace Microsoft.Dafny {
           //   = [Box] F(Ty1, .., TyN, Layer, Heap, self, [Unbox] arg1, .., [Unbox] argN)
 
           var fhandle = FunctionCall(f.tok, name, predef.HandleType, SnocSelf(SnocPrevH(args)));
-          var lhs = FunctionCall(f.tok, Apply(arity), TrType(f.ResultType), Concat(tyargs, Cons(h, Cons(fhandle, lhs_args))));
+          var lhs = FunctionCall(f.tok, Apply(arity), TrType(f.ResultType),
+            Concat(tyargs, Cons(h, Cons(fhandle, lhs_args))));
           var args_h = AlwaysUseHeap || f.ReadsHeap ? Snoc(SnocPrevH(args), h) : args;
           var rhs = FunctionCall(f.tok, f.FullSanitizedName, TrType(f.ResultType), Concat(SnocSelf(args_h), rhs_args));
           var rhs_boxed = BoxIfUnboxed(rhs, f.ResultType);
 
-          sink.AddTopLevelDeclaration(new Axiom(f.tok,
+          AddIncludeDepAxiom(new Axiom(f.tok,
             BplForall(Concat(vars, bvars), BplTrigger(lhs), Bpl.Expr.Eq(lhs, rhs_boxed))));
         }
 
@@ -8678,7 +6792,7 @@ namespace Microsoft.Dafny {
             rhs = FunctionCall(f.tok, RequiresName(f), Bpl.Type.Bool, Concat(SnocSelf(args_h), rhs_args));
           }
 
-          sink.AddTopLevelDeclaration(new Axiom(f.tok,
+          AddIncludeDepAxiom(new Axiom(f.tok,
             BplForall(Concat(vars, bvars), BplTrigger(lhs), Bpl.Expr.Eq(lhs, rhs))));
         }
 
@@ -8711,7 +6825,7 @@ namespace Microsoft.Dafny {
           var rhs_unboxed = UnboxIfBoxed(rhs, f.ResultType);
           var tr = BplTriggerHeap(this, f.tok, lhs, AlwaysUseHeap || f.ReadsHeap ? null : h);
 
-          sink.AddTopLevelDeclaration(new Axiom(f.tok,
+          AddIncludeDepAxiom(new Axiom(f.tok,
             BplForall(Concat(vars, func_vars), tr, Bpl.Expr.Eq(lhs, rhs_unboxed))));
         }
       }
@@ -8866,7 +6980,7 @@ namespace Microsoft.Dafny {
           if (selectorVar == "r") {
             op = (u, v) => Bpl.Expr.Imp(v, u);
           }
-          sink.AddTopLevelDeclaration(new Axiom(tok,
+          AddIncludeDepAxiom(new Axiom(tok,
             BplForall(bvars, BplTrigger(lhs), op(lhs, rhs))));
         };
         SelectorSemantics(Apply(arity), predef.BoxType, "h", apply_ty, Requires(arity), requires_ty);
@@ -9226,78 +7340,29 @@ namespace Microsoft.Dafny {
       Contract.Requires(td != null);
       IToken tok = td.tok;
 
-      var ty_repr = TrType(UserDefinedType.FromTopLevelDecl(td.tok, td));
-      var arity = td.TypeArgs.Count;
-      var inner_name = GetClass(td).TypedIdent.Name;
-      string name = "T" + inner_name;
-      // Create the type constructor
-      if (td is ClassDecl cl && cl.IsObjectTrait) {
-        // the type constructor for "object" is in DafnyPrelude.bpl
-      } else if (td is TupleTypeDecl ttd && ttd.Dims == 2 && ttd.NonGhostDims == 2) {
-        // the type constructor for "Tuple2" is in DafnyPrelude.bpl
-      } else {
-        Bpl.Variable tyVarOut = BplFormalVar(null, predef.Ty, false);
-        List<Bpl.Variable> args = new List<Bpl.Variable>(
-          Enumerable.Range(0, arity).Select(i =>
-            (Bpl.Variable)BplFormalVar(null, predef.Ty, true)));
-        var func = new Bpl.Function(tok, name, args, tyVarOut);
-        sink.AddTopLevelDeclaration(func);
-      }
+      var func = GetOrCreateTypeConstructor(td);
+      var name = func.Name;
 
-      // Helper action to create variables and the function call.
-      Action<Action<List<Bpl.Expr>, List<Bpl.Variable>, Bpl.Expr>> Helper = K => {
-        List<Bpl.Expr> argExprs;
-        var args = MkTyParamBinders(td.TypeArgs, out argExprs);
-        var inner = FunctionCall(tok, name, predef.Ty, argExprs);
-        K(argExprs, args, inner);
-      };
-
-      /* Create the Tag and calling Tag on this type constructor
-       *
-       * The common case:
-       *     const unique TagList: TyTag;
-       *     const unique tytagFamily$List: TyTagFamily;  // defined once for each type named "List"
-       *     axiom (forall t0: Ty :: { List(t0) } Tag(List(t0)) == TagList && TagFamily(List(t0)) == tytagFamily$List);
-       * For types obtained via an abstract import, just do:
-       *     const unique tytagFamily$List: TyTagFamily;  // defined once for each type named "List"
-       *     axiom (forall t0: Ty :: { List(t0) } TagFamily(List(t0)) == tytagFamily$List);
-       */
-      Helper((argExprs, args, inner) => {
-        Bpl.Expr body = Bpl.Expr.True;
-
-        if (!td.EnclosingModuleDefinition.IsFacade) {
-          var tagName = "Tag" + inner_name;
-          var tag = new Bpl.Constant(tok, new Bpl.TypedIdent(tok, tagName, predef.TyTag), true);
-          sink.AddTopLevelDeclaration(tag);
-          body = Bpl.Expr.Eq(FunctionCall(tok, "Tag", predef.TyTag, inner), new Bpl.IdentifierExpr(tok, tag));
-        }
-
-        if (!tytagConstants.TryGetValue(td.Name, out var tagFamily)) {
-          tagFamily = new Bpl.Constant(Token.NoToken, new Bpl.TypedIdent(Token.NoToken, "tytagFamily$" + td.Name, predef.TyTagFamily), true);
-          tytagConstants.Add(td.Name, tagFamily);
-        }
-        body = BplAnd(body, Bpl.Expr.Eq(FunctionCall(tok, "TagFamily", predef.TyTagFamily, inner), new Bpl.IdentifierExpr(tok, tagFamily)));
-
-        var qq = BplForall(args, BplTrigger(inner), body);
-        sink.AddTopLevelDeclaration(new Axiom(tok, qq, name + " Tag"));
-      });
+      var tagAxiom = CreateTagAndCallingForTypeConstructor(td);
+      AddOtherDefinition(func, tagAxiom);
 
       // Create the injectivity axiom and its function
       /*
          function List_0(Ty) : Ty;
          axiom (forall t0: Ty :: { List(t0) } List_0(List(t0)) == t0);
       */
-      for (int i = 0; i < arity; i++) {
-        Helper((argExprs, args, inner) => {
-          Bpl.Variable tyVarIn = BplFormalVar(null, predef.Ty, true);
-          Bpl.Variable tyVarOut = BplFormalVar(null, predef.Ty, false);
-          var injname = name + "_" + i;
-          var injfunc = new Bpl.Function(tok, injname, Singleton(tyVarIn), tyVarOut);
-          var outer = FunctionCall(tok, injname, args[i].TypedIdent.Type, inner);
-          Bpl.Expr qq = BplForall(args, BplTrigger(inner), Bpl.Expr.Eq(outer, argExprs[i]));
-          sink.AddTopLevelDeclaration(new Axiom(tok, qq, name + " injectivity " + i));
-          sink.AddTopLevelDeclaration(injfunc);
-        });
+      for (int i = 0; i < func.InParams.Count; i++) {
+        var args = MkTyParamBinders(td.TypeArgs, out var argExprs);
+        var inner = FunctionCall(tok, name, predef.Ty, argExprs);
+        Bpl.Variable tyVarIn = BplFormalVar(null, predef.Ty, true);
+        Bpl.Variable tyVarOut = BplFormalVar(null, predef.Ty, false);
+        var injname = name + "_" + i;
+        var injfunc = new Bpl.Function(tok, injname, Singleton(tyVarIn), tyVarOut);
+        sink.AddTopLevelDeclaration(injfunc);
+        var outer = FunctionCall(tok, injname, args[i].TypedIdent.Type, inner);
+        Bpl.Expr qq = BplForall(args, BplTrigger(inner), Bpl.Expr.Eq(outer, argExprs[i]));
+        var injectivityAxiom = new Axiom(tok, qq, name + " injectivity " + i);
+        AddOtherDefinition(injfunc, injectivityAxiom);
       }
 
       // Boxing axiom (important for the properties of unbox)
@@ -9309,13 +7374,74 @@ namespace Microsoft.Dafny {
                && $Is($Unbox(bx): DatatypeType, List(T)));
       */
       if (!ModeledAsBoxType(UserDefinedType.FromTopLevelDecl(td.tok, td))) {
-        Helper((argExprs, args, _inner) => {
-          var typeTerm = FunctionCall(tok, name, predef.Ty, argExprs);
-          AddBoxUnboxAxiom(tok, name, typeTerm, ty_repr, args);
-        });
+        var args = MkTyParamBinders(td.TypeArgs, out var argExprs);
+        var ty_repr = TrType(UserDefinedType.FromTopLevelDecl(td.tok, td));
+        var typeTerm = FunctionCall(tok, name, predef.Ty, argExprs);
+        AddBoxUnboxAxiom(tok, name, typeTerm, ty_repr, args);
       }
 
       return name;
+    }
+
+    private Bpl.Function GetOrCreateTypeConstructor(TopLevelDecl td) {
+      Bpl.Function func;
+      if (td is ClassDecl cl && cl.IsObjectTrait) {
+        // the type constructor for "object" is in DafnyPrelude.bpl
+        func = predef.ObjectTypeConstructor;
+      } else if (td is TupleTypeDecl ttd && ttd.Dims == 2 && ttd.NonGhostDims == 2) {
+        // the type constructor for "Tuple2" is in DafnyPrelude.bpl
+        func = this.predef.Tuple2TypeConstructor;
+      } else {
+        var inner_name = GetClass(td).TypedIdent.Name;
+        string name = "T" + inner_name;
+
+        Bpl.Variable tyVarOut = BplFormalVar(null, predef.Ty, false);
+        var args = Enumerable.Range(0, td.TypeArgs.Count).Select(i => (Bpl.Variable)BplFormalVar(null, predef.Ty, true)).ToList();
+        func = new Bpl.Function(td.tok, name, args, tyVarOut);
+        sink.AddTopLevelDeclaration(func);
+      }
+
+      return func;
+    }
+
+    /* Create the Tag and calling Tag on this type constructor
+     *
+     * The common case:
+     *     const unique TagList: TyTag;
+     *     const unique tytagFamily$List: TyTagFamily;  // defined once for each type named "List"
+     *     axiom (forall t0: Ty :: { List(t0) } Tag(List(t0)) == TagList && TagFamily(List(t0)) == tytagFamily$List);
+     * For types obtained via an abstract import, just do:
+     *     const unique tytagFamily$List: TyTagFamily;  // defined once for each type named "List"
+     *     axiom (forall t0: Ty :: { List(t0) } TagFamily(List(t0)) == tytagFamily$List);
+     */
+    private Axiom CreateTagAndCallingForTypeConstructor(TopLevelDecl td) {
+      IToken tok = td.tok;
+      var inner_name = GetClass(td).TypedIdent.Name;
+      string name = "T" + inner_name;
+
+      var args = MkTyParamBinders(td.TypeArgs, out var argExprs);
+      var inner = FunctionCall(tok, name, predef.Ty, argExprs);
+      Bpl.Expr body = Bpl.Expr.True;
+
+      if (!td.EnclosingModuleDefinition.IsFacade) {
+        var tagName = "Tag" + inner_name;
+        var tag = new Bpl.Constant(tok, new Bpl.TypedIdent(tok, tagName, predef.TyTag), true);
+        sink.AddTopLevelDeclaration(tag);
+        body = Bpl.Expr.Eq(FunctionCall(tok, "Tag", predef.TyTag, inner), new Bpl.IdentifierExpr(tok, tag));
+      }
+
+      if (!tytagConstants.TryGetValue(td.Name, out var tagFamily)) {
+        tagFamily = new Bpl.Constant(Token.NoToken,
+          new Bpl.TypedIdent(Token.NoToken, "tytagFamily$" + td.Name, predef.TyTagFamily), true);
+        tytagConstants.Add(td.Name, tagFamily);
+      }
+
+      body = BplAnd(body,
+        Bpl.Expr.Eq(FunctionCall(tok, "TagFamily", predef.TyTagFamily, inner), new Bpl.IdentifierExpr(tok, tagFamily)));
+
+      var qq = BplForall(args, BplTrigger(inner), body);
+      var tagAxiom = new Axiom(tok, qq, name + " Tag");
+      return tagAxiom;
     }
 
     /// <summary>
@@ -9402,7 +7528,7 @@ namespace Microsoft.Dafny {
         var ig = FunctionCall(f.tok, BuiltinFunction.IsGhostField, ty, Bpl.Expr.Ident(fc));
         cond = Bpl.Expr.And(cond, f.IsGhost ? ig : Bpl.Expr.Not(ig));
         Bpl.Axiom ax = new Bpl.Axiom(f.tok, cond);
-        sink.AddTopLevelDeclaration(ax);
+        AddIncludeDepAxiom(ax);
       }
       return fc;
     }
@@ -9510,12 +7636,13 @@ namespace Microsoft.Dafny {
     /// <summary>
     /// This method is expected to be called just once for each function in the program.
     /// </summary>
-    void AddFunction(Function f) {
+    Bpl.Function AddFunction(Function f) {
       Contract.Requires(f != null);
       Contract.Requires(predef != null && sink != null);
 
       // declare the function
-      if (!f.IsBuiltin) {
+      Bpl.Function func;
+      {
         var formals = new List<Variable>();
         formals.AddRange(MkTyParamFormals(GetTypeParams(f)));
         if (f.IsFuelAware()) {
@@ -9534,7 +7661,7 @@ namespace Microsoft.Dafny {
           formals.Add(new Bpl.Formal(p.tok, new Bpl.TypedIdent(p.tok, p.AssignUniqueName(f.IdGenerator), TrType(p.Type)), true));
         }
         var res = new Bpl.Formal(f.tok, new Bpl.TypedIdent(f.tok, Bpl.TypedIdent.NoName, TrType(f.ResultType)), false);
-        var func = new Bpl.Function(f.tok, f.FullSanitizedName, new List<Bpl.TypeVariable>(), formals, res, "function declaration for " + f.FullName);
+        func = new Bpl.Function(f.tok, f.FullSanitizedName, new List<Bpl.TypeVariable>(), formals, res, "function declaration for " + f.FullName);
         if (InsertChecksums) {
           InsertChecksum(f, func);
         }
@@ -9561,6 +7688,8 @@ namespace Microsoft.Dafny {
         var canCallF = new Bpl.Function(f.tok, f.FullSanitizedName + "#canCall", new List<Bpl.TypeVariable>(), formals, res);
         sink.AddTopLevelDeclaration(canCallF);
       }
+
+      return func;
     }
 
     /// <summary>
@@ -9592,165 +7721,6 @@ namespace Microsoft.Dafny {
     /// but no callers, and vice versa for InterModuleCall, IntraModuleCall, and CoCall.
     /// </summary>
     enum MethodTranslationKind { SpecWellformedness, Call, CoCall, Implementation, OverrideCheck }
-
-    /// <summary>
-    /// This method is expected to be called at most once for each parameter combination, and in particular
-    /// at most once for each value of "kind".
-    /// </summary>
-    Bpl.Procedure AddMethod(Method m, MethodTranslationKind kind) {
-      Contract.Requires(m != null);
-      Contract.Requires(m.EnclosingClass != null);
-      Contract.Requires(predef != null);
-      Contract.Requires(currentModule == null && codeContext == null && isAllocContext == null);
-      Contract.Ensures(currentModule == null && codeContext == null && isAllocContext == null);
-      Contract.Ensures(Contract.Result<Bpl.Procedure>() != null);
-      Contract.Assert(VisibleInScope(m));
-
-      currentModule = m.EnclosingClass.EnclosingModuleDefinition;
-      codeContext = m;
-      isAllocContext = new IsAllocContext(m.IsGhost);
-
-      Bpl.Expr prevHeap = null;
-      Bpl.Expr currHeap = null;
-      var ordinaryEtran = new ExpressionTranslator(this, predef, m.tok);
-      ExpressionTranslator etran;
-      var inParams = new List<Bpl.Variable>();
-      if (m is TwoStateLemma) {
-        var prevHeapVar = new Bpl.Formal(m.tok, new Bpl.TypedIdent(m.tok, "previous$Heap", predef.HeapType), true);
-        var currHeapVar = new Bpl.Formal(m.tok, new Bpl.TypedIdent(m.tok, "current$Heap", predef.HeapType), true);
-        inParams.Add(prevHeapVar);
-        inParams.Add(currHeapVar);
-        prevHeap = new Bpl.IdentifierExpr(m.tok, prevHeapVar);
-        currHeap = new Bpl.IdentifierExpr(m.tok, currHeapVar);
-        etran = new ExpressionTranslator(this, predef, currHeap, prevHeap);
-      } else {
-        etran = ordinaryEtran;
-      }
-
-      List<Variable> outParams;
-      GenerateMethodParameters(m.tok, m, kind, etran, inParams, out outParams);
-
-      var req = new List<Bpl.Requires>();
-      var mod = new List<Bpl.IdentifierExpr>();
-      var ens = new List<Bpl.Ensures>();
-      // FREE PRECONDITIONS
-      if (kind == MethodTranslationKind.SpecWellformedness || kind == MethodTranslationKind.Implementation || kind == MethodTranslationKind.OverrideCheck) {  // the other cases have no need for a free precondition
-        // free requires mh == ModuleContextHeight && fh == FunctionContextHeight;
-        req.Add(Requires(m.tok, true, etran.HeightContext(kind == MethodTranslationKind.OverrideCheck ? m.OverriddenMethod : m), null, null));
-        if (m is TwoStateLemma) {
-          // free requires prevHeap == Heap && HeapSucc(prevHeap, currHeap) && IsHeap(currHeap)
-          var a0 = Bpl.Expr.Eq(prevHeap, ordinaryEtran.HeapExpr);
-          var a1 = HeapSucc(prevHeap, currHeap);
-          var a2 = FunctionCall(m.tok, BuiltinFunction.IsGoodHeap, null, currHeap);
-          req.Add(Requires(m.tok, true, BplAnd(a0, BplAnd(a1, a2)), null, null));
-        }
-      }
-      if (m is TwoStateLemma) {
-        // Checked preconditions that old parameters really existed in previous state
-        var index = 0;
-        foreach (var formal in m.Ins) {
-          if (formal.IsOld) {
-            var dafnyFormalIdExpr = new IdentifierExpr(formal.tok, formal);
-            req.Add(Requires(formal.tok, false, MkIsAlloc(etran.TrExpr(dafnyFormalIdExpr), formal.Type, prevHeap),
-              string.Format("parameter{0} ('{1}') must be allocated in the two-state lemma's previous state",
-              m.Ins.Count == 1 ? "" : " " + index, formal.Name), null));
-          }
-          index++;
-        }
-      }
-      mod.Add((Bpl.IdentifierExpr/*TODO: this cast is somewhat dubious*/)ordinaryEtran.HeapExpr);
-      mod.Add(etran.Tick());
-
-      var bodyKind = kind == MethodTranslationKind.SpecWellformedness || kind == MethodTranslationKind.Implementation;
-
-      if (kind != MethodTranslationKind.SpecWellformedness && kind != MethodTranslationKind.OverrideCheck) {
-        // USER-DEFINED SPECIFICATIONS
-        var comment = "user-defined preconditions";
-        foreach (var p in m.Req) {
-          string errorMessage = CustomErrorMessage(p.Attributes);
-          if (p.Label != null && kind == MethodTranslationKind.Implementation) {
-            // don't include this precondition here, but record it for later use
-            p.Label.E = (m is TwoStateLemma ? ordinaryEtran : etran.Old).TrExpr(p.E);
-          } else {
-            foreach (var s in TrSplitExprForMethodSpec(p.E, etran, kind)) {
-              if (s.IsOnlyChecked && bodyKind) {
-                // don't include in split
-              } else if (s.IsOnlyFree && !bodyKind) {
-                // don't include in split -- it would be ignored, anyhow
-              } else {
-                req.Add(Requires(s.E.tok, s.IsOnlyFree, s.E, errorMessage, comment));
-                comment = null;
-                // the free here is not linked to the free on the original expression (this is free things generated in the splitting.)
-              }
-            }
-          }
-        }
-        comment = "user-defined postconditions";
-        foreach (var p in m.Ens) {
-          string errorMessage = CustomErrorMessage(p.Attributes);
-          AddEnsures(ens, Ensures(p.E.tok, true, CanCallAssumption(p.E, etran), errorMessage, comment));
-          comment = null;
-          foreach (var s in TrSplitExprForMethodSpec(p.E, etran, kind)) {
-            var post = s.E;
-            if (kind == MethodTranslationKind.Implementation && RefinementToken.IsInherited(s.E.tok, currentModule)) {
-              // this postcondition was inherited into this module, so make it into the form "$_reverifyPost ==> s.E"
-              post = Bpl.Expr.Imp(new Bpl.IdentifierExpr(s.E.tok, "$_reverifyPost", Bpl.Type.Bool), post);
-            }
-            if (s.IsOnlyFree && bodyKind) {
-              // don't include in split -- it would be ignored, anyhow
-            } else if (s.IsOnlyChecked && !bodyKind) {
-              // don't include in split
-            } else {
-              AddEnsures(ens, Ensures(s.E.tok, s.IsOnlyFree, post, errorMessage, null));
-            }
-          }
-        }
-        if (m is Constructor && kind == MethodTranslationKind.Call) {
-          var fresh = Bpl.Expr.Not(etran.Old.IsAlloced(m.tok, new Bpl.IdentifierExpr(m.tok, "this", TrReceiverType(m))));
-          AddEnsures(ens, Ensures(m.tok, false, fresh, null, "constructor allocates the object"));
-        }
-        foreach (BoilerplateTriple tri in GetTwoStateBoilerplate(m.tok, m.Mod.Expressions, m.IsGhost, m.AllowsAllocation, ordinaryEtran.Old, ordinaryEtran, ordinaryEtran.Old)) {
-          AddEnsures(ens, Ensures(tri.tok, tri.IsFree, tri.Expr, tri.ErrorMessage, tri.Comment));
-        }
-
-        // add the fuel assumption for the reveal method of a opaque method
-        if (IsOpaqueRevealLemma(m)) {
-          List<Expression> args = Attributes.FindExpressions(m.Attributes, "fuel");
-          if (args != null) {
-            MemberSelectExpr selectExpr = args[0].Resolved as MemberSelectExpr;
-            if (selectExpr != null) {
-              Function f = selectExpr.Member as Function;
-              FuelConstant fuelConstant = this.functionFuel.Find(x => x.f == f);
-              if (fuelConstant != null) {
-                Bpl.Expr startFuel = fuelConstant.startFuel;
-                Bpl.Expr startFuelAssert = fuelConstant.startFuelAssert;
-                Bpl.Expr moreFuel_expr = fuelConstant.MoreFuel(sink, predef, f.IdGenerator);
-                Bpl.Expr layer = etran.layerInterCluster.LayerN(1, moreFuel_expr);
-                Bpl.Expr layerAssert = etran.layerInterCluster.LayerN(2, moreFuel_expr);
-
-                AddEnsures(ens, Ensures(m.tok, true, Bpl.Expr.Eq(startFuel, layer), null, null));
-                AddEnsures(ens, Ensures(m.tok, true, Bpl.Expr.Eq(startFuelAssert, layerAssert), null, null));
-
-                AddEnsures(ens, Ensures(m.tok, true, Bpl.Expr.Eq(FunctionCall(f.tok, BuiltinFunction.AsFuelBottom, null, moreFuel_expr), moreFuel_expr), null, "Shortcut to LZ"));
-              }
-            }
-          }
-        }
-      }
-
-      var name = MethodName(m, kind);
-      var proc = new Bpl.Procedure(m.tok, name, new List<Bpl.TypeVariable>(), inParams, outParams, req, mod, ens, etran.TrAttributes(m.Attributes, null));
-
-      if (InsertChecksums) {
-        InsertChecksum(m, proc, true);
-      }
-
-      currentModule = null;
-      codeContext = null;
-      isAllocContext = null;
-
-      return proc;
-    }
 
     static string MethodName(ICodeContext m, MethodTranslationKind kind) {
       Contract.Requires(m != null);
@@ -12221,7 +10191,7 @@ namespace Microsoft.Dafny {
             var p = Substitute(e.RHSs[0], receiverReplacement, substMap);
             Bpl.Expr ax = Bpl.Expr.Imp(canCall, BplAnd(antecedent, etranCC.TrExpr(p)));
             ax = BplForall(gg, tr, ax);
-            sink.AddTopLevelDeclaration(new Bpl.Axiom(e.tok, ax));
+            AddIncludeDepAxiom(new Bpl.Axiom(e.tok, ax));
           }
 
           // now that we've declared the functions and axioms, let's prepare the let-such-that desugaring
