+++ Source/DafnyPipeline.Test/IntraMethodVerificationStability.cs
@@ -1,55 +1,153 @@
-using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Microsoft.Boogie;
 using Microsoft.Dafny;
 using Xunit;
+using Xunit.Abstractions;
 using BoogieProgram = Microsoft.Boogie.Program;
 using Parser = Microsoft.Dafny.Parser;
-using Type = System.Type;
 
 namespace DafnyPipeline.Test {
   // Main.Resolve has static shared state (TypeConstraint.ErrorsToBeReported for example)
   // so we can't execute tests that use it in parallel.
   [Collection("Singleton Test Collection - Resolution")]
   public class IntraMethodVerificationStability {
-    [Fact]
-    public void NoUniqueLinesWhenConcatenatingUnrelatedPrograms() {
-      var program = $@"
+    private readonly ITestOutputHelper testOutputHelper;
+
+    // All types of top level declarations.
+    readonly string originalProgram = @"
+module SomeModule {
+
+  module NestedModule {
+    class C {
+      var f: int
+      constructor ()
+    }
+  }
+
+  method m() {
+    var x: NestedModule.C;
+    x := new NestedModule.C();
+    x.f := 4;
+  }
+}
+
+import opened SomeModule
+
+type FooSynonym<T> = FooClass
+
+class FooClass {
+  var f: int
+  constructor ()
+}
+
 datatype Friends = Agnes | Agatha | Jermaine
 
-function SomeFunc(funcFormal: int): nat {{ 3 }}
+function method SomeFunc(funcFormal: int): nat { 3 }
 
 method SomeMethod(methodFormal: int) returns (result: bool)
+  requires methodFormal == 2
   ensures result == true 
-{{
-  result := methodFormal == 3;
-}}
+  // ensures forall x :: x == methodFormal
+{
+  m();
+  var lambdaExpr := x => x + 1;
+  result := methodFormal == SomeFunc(42);
+}
 ";
 
-      var renamedProgram = $@"
+    readonly string renamedProgram = @"   
+
+module SomeModule2 {
+
+  module NestedModule {
+      class C {
+        var f: int
+        constructor ()
+      }
+    }
+
+    method m() {
+      var x: NestedModule.C;
+      x := new NestedModule.C();
+      x.f := 4;
+    }
+}
+
+type FooSynonym2<T> = FooClass2
+
+class FooClass2 {
+  var f: int
+  constructor ()
+}
+
 datatype Friends2 = Agnes2 | Agatha2 | Jermaine2
 
-function SomeFunc2(funcFormal: int): nat {{ 3 }}
+function method SomeFunc2(funcFormal: int): nat { 3 }
 
 method SomeMethod2(methodFormal: int) returns (result: bool) 
+  requires methodFormal == 2
   ensures result == true
-{{
-  result := methodFormal == 3;
-}}
+  // ensures forall x :: x == methodFormal
+{
+  var lambdaExpr := x => x + 1;
+  result := methodFormal == SomeFunc2(42);
+}
 ";
 
+    public IntraMethodVerificationStability(ITestOutputHelper testOutputHelper) {
+      this.testOutputHelper = testOutputHelper;
+    }
+
+    [Fact]
+    public void NoUniqueLinesWhenConcatenatingUnrelatedPrograms() {
       DafnyOptions.Install(new DafnyOptions());
 
-      var regularBoogie = GetBoogieText(program);
-      var renamedBoogie = GetBoogieText(renamedProgram);
-      var separate = UniqueNonCommentLines(regularBoogie + renamedBoogie);
-      var combinedBoogie = GetBoogieText(program + renamedProgram);
+      var regularBoogie = GetBoogie(originalProgram).ToList();
+      var renamedBoogie = GetBoogie(renamedProgram).ToList();
+      var regularBoogieText = GetBoogieText(regularBoogie);
+      var renamedBoogieText = GetBoogieText(renamedBoogie);
+      var separate = UniqueNonCommentLines(regularBoogieText + renamedBoogieText);
+      var combinedBoogie = GetBoogieText(GetBoogie(originalProgram + renamedProgram));
       var together = UniqueNonCommentLines(combinedBoogie);
 
       var uniqueLines = separate.Union(together).Except(separate.Intersect(together)).ToList();
       Assert.Equal(Enumerable.Empty<string>(), uniqueLines);
+    }
+
+    [Fact]
+    public void EqualProverLogWhenAddingUnrelatedProgram() {
+
+      DafnyOptions.Install(new DafnyOptions());
+      CommandLineOptions.Clo.Parse(new[] { "" });
+      CommandLineOptions.Clo.ProcsToCheck.Add("*SomeMethod");
+      ExecutionEngine.printer = new ConsolePrinter(); // For boogie errors
+
+      var renamedProverLog = GetProverLogForProgram(GetBoogie(renamedProgram + originalProgram));
+      var regularProverLog = GetProverLogForProgram(GetBoogie(originalProgram));
+      Assert.Equal(regularProverLog, renamedProverLog);
+    }
+
+    private string GetProverLogForProgram(IEnumerable<Microsoft.Boogie.Program> boogiePrograms) {
+      var logs = GetProverLogsForProgram(boogiePrograms).ToList();
+      Assert.Single(logs);
+      return logs[0];
+    }
+
+    private IEnumerable<string> GetProverLogsForProgram(IEnumerable<Microsoft.Boogie.Program> boogiePrograms) {
+      string directory = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
+      Directory.CreateDirectory(directory);
+      var temp1 = directory + "/proverLog";
+      testOutputHelper.WriteLine("proverLog: " + temp1);
+      CommandLineOptions.Clo.ProverLogFilePath = temp1;
+      foreach (var boogieProgram in boogiePrograms) {
+        Main.BoogieOnce("", "", boogieProgram, "programId", out _, out var outcome);
+        testOutputHelper.WriteLine("outcome: " + outcome);
+        foreach (var proverFile in Directory.GetFiles(directory)) {
+          yield return File.ReadAllText(proverFile);
+        }
+      }
     }
 
     ISet<string> UniqueNonCommentLines(string input) {
@@ -63,8 +161,8 @@ method SomeMethod2(methodFormal: int) returns (result: bool)
       return result.ToString();
     }
 
-    string GetBoogieText(string dafnyProgramText) {
-      return string.Join('\n', GetBoogie(dafnyProgramText).Select(PrintBoogie));
+    string GetBoogieText(IEnumerable<BoogieProgram> boogieProgram) {
+      return string.Join('\n', boogieProgram.Select(PrintBoogie));
     }
 
     IEnumerable<BoogieProgram> GetBoogie(string dafnyProgramText) {
@@ -77,8 +175,8 @@ method SomeMethod2(methodFormal: int) returns (result: bool)
       Assert.Equal(0, parseResult);
       var dafnyProgram = new Microsoft.Dafny.Program(fullFilePath, module, builtIns, errorReporter);
       Main.Resolve(dafnyProgram, errorReporter);
-      var boogiePrograms = Translator.Translate(dafnyProgram, errorReporter).Select(t => t.Item2);
-      return boogiePrograms;
+      Assert.Equal(0, errorReporter.ErrorCount);
+      return Translator.Translate(dafnyProgram, errorReporter).Select(t => t.Item2).ToList();
     }
   }
