+++ Source/Directory.Build.props
@@ -7,7 +7,7 @@
 
   <!-- Boogie dependency -->
   <ItemGroup>
-    <PackageReference Include="Boogie.ExecutionEngine" Version="2.9.6" />
+    <PackageReference Include="Boogie.ExecutionEngine" Version="2.11.0" />
   </ItemGroup>
 
 </Project>
